% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPairwiseMatrix.R
\name{AnnotationPairwiseMatrix}
\alias{AnnotationPairwiseMatrix}
\title{Single-cell Annotation Pairwise Matrix}
\description{
Single-cell Annotation Pairwise Matrix

Single-cell Annotation Pairwise Matrix
}
\details{
Base class for matrix-like data storing pairwise relationships with rows and
columns aligned to either the observations or features of \code{X}
(\code{\link{AssayMatrix}}).
}
\section{Super classes}{
\code{\link[tiledbsc:TileDBObject]{tiledbsc::TileDBObject}} -> \code{\link[tiledbsc:TileDBArray]{tiledbsc::TileDBArray}} -> \code{\link[tiledbsc:AnnotationArray]{tiledbsc::AnnotationArray}} -> \code{AnnotationPairwiseMatrix}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnotationPairwiseMatrix-from_matrix}{\code{AnnotationPairwiseMatrix$from_matrix()}}
\item \href{#method-AnnotationPairwiseMatrix-to_matrix}{\code{AnnotationPairwiseMatrix$to_matrix()}}
\item \href{#method-AnnotationPairwiseMatrix-to_dataframe}{\code{AnnotationPairwiseMatrix$to_dataframe()}}
\item \href{#method-AnnotationPairwiseMatrix-to_sparse_matrix}{\code{AnnotationPairwiseMatrix$to_sparse_matrix()}}
\item \href{#method-AnnotationPairwiseMatrix-to_seurat_graph}{\code{AnnotationPairwiseMatrix$to_seurat_graph()}}
\item \href{#method-AnnotationPairwiseMatrix-clone}{\code{AnnotationPairwiseMatrix$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBObject" data-id="class"><a href='../../tiledbsc/html/TileDBObject.html#method-TileDBObject-class'><code>tiledbsc::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBObject" data-id="exists"><a href='../../tiledbsc/html/TileDBObject.html#method-TileDBObject-exists'><code>tiledbsc::TileDBObject$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="add_metadata"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-add_metadata'><code>tiledbsc::TileDBArray$add_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="array_exists"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-array_exists'><code>tiledbsc::TileDBArray$array_exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="attributes"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-attributes'><code>tiledbsc::TileDBArray$attributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="attrnames"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-attrnames'><code>tiledbsc::TileDBArray$attrnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="dimensions"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-dimensions'><code>tiledbsc::TileDBArray$dimensions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="dimnames"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-dimnames'><code>tiledbsc::TileDBArray$dimnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="fragment_count"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-fragment_count'><code>tiledbsc::TileDBArray$fragment_count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="get_metadata"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-get_metadata'><code>tiledbsc::TileDBArray$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="initialize"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-initialize'><code>tiledbsc::TileDBArray$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="print"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-print'><code>tiledbsc::TileDBArray$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="reset_query"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-reset_query'><code>tiledbsc::TileDBArray$reset_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="schema"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-schema'><code>tiledbsc::TileDBArray$schema()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="set_query"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-set_query'><code>tiledbsc::TileDBArray$set_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="tiledb_array"><a href='../../tiledbsc/html/TileDBArray.html#method-TileDBArray-tiledb_array'><code>tiledbsc::TileDBArray$tiledb_array()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationPairwiseMatrix-from_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationPairwiseMatrix-from_matrix}{}}}
\subsection{Method \code{from_matrix()}}{
Ingest annotation matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrix$from_matrix(x, index_cols, value_col = "value")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{any \code{matrix}-like object coercible to a
\code{\link[Matrix:TsparseMatrix-class]{TsparseMatrix}} with string dimensions.}

\item{\code{index_cols}}{Names to use for the TileDB array's dimensions that
will contain the matrix row/column names.}

\item{\code{value_col}}{Name to use for the TileDB array's attribute that will
contain the matrix values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationPairwiseMatrix-to_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationPairwiseMatrix-to_matrix}{}}}
\subsection{Method \code{to_matrix()}}{
Read annotation data from TileDB into a matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrix$to_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link{matrix}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationPairwiseMatrix-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationPairwiseMatrix-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Read annotation data from TileDB into a data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrix$to_dataframe(attrs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attrs}}{Specify one or more attributes to retrieve. If \code{NULL},
all attributes are retrieved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{data.frame}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationPairwiseMatrix-to_sparse_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationPairwiseMatrix-to_sparse_matrix}{}}}
\subsection{Method \code{to_sparse_matrix()}}{
Read annotation data from TileDB into a sparse matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrix$to_sparse_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[Matrix:dgTMatrix-class]{Matrix::dgTMatrix}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationPairwiseMatrix-to_seurat_graph"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationPairwiseMatrix-to_seurat_graph}{}}}
\subsection{Method \code{to_seurat_graph()}}{
Read annotation data from TileDB into Seurat Graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrix$to_seurat_graph()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[SeuratObject:Graph-class]{SeuratObject::Graph}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationPairwiseMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationPairwiseMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
