% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPairwiseMatrixGroup.R
\name{AnnotationPairwiseMatrixGroup}
\alias{AnnotationPairwiseMatrixGroup}
\title{Single-cell Annotation Matrix Group}
\description{
Class for representing a TileDB group containing one or more
\code{\link{AnnotationPairwiseMatrix}} arrays that share a common dimension name.
}
\section{Super classes}{
\code{\link[tiledbsc:TileDBObject]{tiledbsc::TileDBObject}} -> \code{\link[tiledbsc:TileDBGroup]{tiledbsc::TileDBGroup}} -> \code{\link[tiledbsc:AnnotationGroup]{tiledbsc::AnnotationGroup}} -> \code{AnnotationPairwiseMatrixGroup}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnotationPairwiseMatrixGroup-add_matrix}{\code{AnnotationPairwiseMatrixGroup$add_matrix()}}
\item \href{#method-AnnotationPairwiseMatrixGroup-add_seurat_graph}{\code{AnnotationPairwiseMatrixGroup$add_seurat_graph()}}
\item \href{#method-AnnotationPairwiseMatrixGroup-clone}{\code{AnnotationPairwiseMatrixGroup$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBObject" data-id="class"><a href='../../tiledbsc/html/TileDBObject.html#method-TileDBObject-class'><code>tiledbsc::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBObject" data-id="exists"><a href='../../tiledbsc/html/TileDBObject.html#method-TileDBObject-exists'><code>tiledbsc::TileDBObject$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_member"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-add_member'><code>tiledbsc::TileDBGroup$add_member()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_metadata"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-add_metadata'><code>tiledbsc::TileDBGroup$add_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="count_members"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-count_members'><code>tiledbsc::TileDBGroup$count_members()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_member"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-get_member'><code>tiledbsc::TileDBGroup$get_member()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_members"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-get_members'><code>tiledbsc::TileDBGroup$get_members()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_metadata"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-get_metadata'><code>tiledbsc::TileDBGroup$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="group_exists"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-group_exists'><code>tiledbsc::TileDBGroup$group_exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_member_uris"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-list_member_uris'><code>tiledbsc::TileDBGroup$list_member_uris()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_members"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-list_members'><code>tiledbsc::TileDBGroup$list_members()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_object_uris"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-list_object_uris'><code>tiledbsc::TileDBGroup$list_object_uris()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_objects"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-list_objects'><code>tiledbsc::TileDBGroup$list_objects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="print"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-print'><code>tiledbsc::TileDBGroup$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="remove_member"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-remove_member'><code>tiledbsc::TileDBGroup$remove_member()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="tiledb_group"><a href='../../tiledbsc/html/TileDBGroup.html#method-TileDBGroup-tiledb_group'><code>tiledbsc::TileDBGroup$tiledb_group()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="AnnotationGroup" data-id="initialize"><a href='../../tiledbsc/html/AnnotationGroup.html#method-AnnotationGroup-initialize'><code>tiledbsc::AnnotationGroup$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="AnnotationGroup" data-id="reset_query"><a href='../../tiledbsc/html/AnnotationGroup.html#method-AnnotationGroup-reset_query'><code>tiledbsc::AnnotationGroup$reset_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="AnnotationGroup" data-id="set_query"><a href='../../tiledbsc/html/AnnotationGroup.html#method-AnnotationGroup-set_query'><code>tiledbsc::AnnotationGroup$set_query()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationPairwiseMatrixGroup-add_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationPairwiseMatrixGroup-add_matrix}{}}}
\subsection{Method \code{add_matrix()}}{
Add a new \code{\link{AnnotationPairwiseMatrix}} array to the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrixGroup$add_matrix(data, name, metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a \code{\link{matrix}} of annotation data to ingest. The \code{matrix} rows
must be aligned to the dimension indicated by the group's
\code{dimension_name}.}

\item{\code{name}}{Name of the new pairwise annotation matrix.}

\item{\code{metadata}}{Named list of metadata to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationPairwiseMatrixGroup-add_seurat_graph"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationPairwiseMatrixGroup-add_seurat_graph}{}}}
\subsection{Method \code{add_seurat_graph()}}{
Convert a \code{\link[SeuratObject:Graph-class]{SeuratObject::Graph}} object to
\code{\link{AnnotationPairwiseMatrix}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrixGroup$add_seurat_graph(object, technique)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{\link[SeuratObject:Graph-class]{SeuratObject::Graph}} object.}

\item{\code{technique}}{Name of the technique used to generate the graph
(typically, \code{nn} or \code{snn}).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\subsection{On-Disk Format}{

Arrays are named \verb{graph_<technique>}.
}

\subsection{Metadata}{
\itemize{
\item \code{assay_used}: Name of the assay used to generate the graph.
\item \code{graph_technique}: Name of the technique used to generate the graph.
used.
}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationPairwiseMatrixGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationPairwiseMatrixGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrixGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
