% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-readerTransformers.R
\name{arrow_table_to_dense}
\alias{arrow_table_to_dense}
\title{Transformer function: Arrow table to matrix}
\usage{
arrow_table_to_dense(tbl, byrow)
}
\arguments{
\item{tbl}{\link[arrow]{Table} with columns "soma_dim_0", "soma_dim_1", and "soma_datah"}

\item{byrow}{Logical, TRUE if "soma_data" is ordered by row, this argument is directly passed
to the argument \code{byrow} of \link{matrix}}
}
\value{
\link{matrixZeroBasedView} of \link[base]{matrix}
}
\description{
Converts a \link[arrow]{Table} of sparse format (columns: "soma_dim_0",
"soma_dim_1", "soma_data") to a \link{matrixZeroBasedView} of a \link{matrix}.
}
