% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coercions.R
\name{as.data.frame.SOMASparseNDArrayRead}
\alias{as.data.frame.SOMASparseNDArrayRead}
\title{Coerce \link[tiledbsoma]{SOMASparseNDArrayRead} to \link{data.frame} or \link[tibble]{tibble}}
\usage{
\method{as.data.frame}{SOMASparseNDArrayRead}(x, ...)
}
\description{
Coerce \link[tiledbsoma]{SOMASparseNDArrayRead} to \link{data.frame} or \link[tibble]{tibble}
}
