% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{as.one.based.matrixZeroBasedView}
\alias{as.one.based.matrixZeroBasedView}
\title{Get one-based matrix object underlying zero-based view}
\usage{
\method{as.one.based}{matrixZeroBasedView}(x)
}
\arguments{
\item{x}{The zero-based matrix view.}
}
\value{
The one-based \code{matrix} or \code{Matrix::sparseMatrix} underlying the view.
}
\description{
Get one-based matrix object underlying zero-based view
}
