% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{as.one.based}
\alias{as.one.based}
\title{Get one-based object}
\usage{
as.one.based(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A one-based version/view of the object.
}
\description{
Get one-based object
}
