% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coercions.R
\name{as_arrow_table.SOMASparseNDArrayRead}
\alias{as_arrow_table.SOMASparseNDArrayRead}
\title{Coercion methods for SOMA classes}
\usage{
\method{as_arrow_table}{SOMASparseNDArrayRead}(x)
}
\description{
Coercion methods for SOMA classes
}
