% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockwiseIter.R
\name{BlockwiseReadIterBase}
\alias{BlockwiseReadIterBase}
\title{SOMA Blockwise Read Iterator Base Class}
\description{
Virtual base class that allows for blockwise read iteration of
SOMA reads.
}
\seealso{
Derived classes: \code{\link{BlockwiseTableReadIter}},
\code{\link{BlockwiseSparseReadIter}}
}
\keyword{internal}
\section{Super class}{
\code{\link[tiledbsoma:ReadIter]{ReadIter}} -> \code{BlockwiseReadIterBase}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{array}}{The underlying SOMA array.}

    \item{\code{axis}}{The axis to iterate over in a blockwise fashion.}

    \item{\code{axes_to_reindex}}{The axes to re-index.}

    \item{\code{coords}}{A list of \code{\link{CoordsStrider}} objects.}

    \item{\code{coords_axis}}{The \code{\link{CoordsStrider}} for \code{axis}.}

    \item{\code{reindex_disable_on_axis}}{Additional axes that will not be
re-indexed.}

    \item{\code{reindexable}}{Shorthand to see if this iterator is poised to be
re-indexed or not.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-BlockwiseReadIterBase-initialize}{\code{BlockwiseReadIterBase$new()}}
    \item \href{#method-BlockwiseReadIterBase-read_complete}{\code{BlockwiseReadIterBase$read_complete()}}
    \item \href{#method-BlockwiseReadIterBase-read_next}{\code{BlockwiseReadIterBase$read_next()}}
    \item \href{#method-BlockwiseReadIterBase-clone}{\code{BlockwiseReadIterBase$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="concat"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-concat'><code>ReadIter$concat()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockwiseReadIterBase-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-BlockwiseReadIterBase-initialize}{}}}
\subsection{\code{BlockwiseReadIterBase$new()}}{
  Create.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BlockwiseReadIterBase$new(
  sr,
  array,
  coords,
  axis,
  ...,
  reindex_disable_on_axis = NA
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{sr}}{SOMA read pointer}
      \item{\code{array}}{Underlying \code{\link{SOMASparseNDArray}}}
      \item{\code{coords}}{Named list of \code{\link{CoordsStrider}} objects; must be
named after \code{array$dimnames()}}
      \item{\code{axis}}{Axis to iterate over in a blockwise manner}
      \item{\code{...}}{Ignored}
      \item{\code{reindex_disable_on_axis}}{Additional axes that will not be re-indexed;
the following values may be used as shorthands for common settings:
\itemize{
\item \dQuote{\code{TRUE}}: disable re-indexing on all axes
\item \dQuote{\code{NA}}: re-index only on \code{axis}, disable
re-indexing on all others
\item \dQuote{\code{FALSE}}: re-index on \emph{all} axes, do \strong{not}
disable re-indexing
}}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockwiseReadIterBase-read_complete"></a>}}
\if{latex}{\out{\hypertarget{method-BlockwiseReadIterBase-read_complete}{}}}
\subsection{\code{BlockwiseReadIterBase$read_complete()}}{
  Check if the iterated read is complete or not.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BlockwiseReadIterBase$read_complete()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{TRUE} if read is complete, otherwise \code{FALSE}.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockwiseReadIterBase-read_next"></a>}}
\if{latex}{\out{\hypertarget{method-BlockwiseReadIterBase-read_next}{}}}
\subsection{\code{BlockwiseReadIterBase$read_next()}}{
  Read the next chunk of the iterated read. If read
is complete, throws an \code{iterationCompleteWarning} warning and
returns \code{NULL}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BlockwiseReadIterBase$read_next()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} or the next blockwise chunk of the iterated read.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockwiseReadIterBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlockwiseReadIterBase-clone}{}}}
\subsection{\code{BlockwiseReadIterBase$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BlockwiseReadIterBase$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
