% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockwiseIter.R
\name{BlockwiseSparseReadIter}
\alias{BlockwiseSparseReadIter}
\title{SOMA Blockwise Read Iterator for Sparse Matrices}
\description{
...
}
\keyword{internal}
\section{Super classes}{
\code{\link[tiledbsoma:ReadIter]{tiledbsoma::ReadIter}} -> \code{\link[tiledbsoma:BlockwiseReadIterBase]{tiledbsoma::BlockwiseReadIterBase}} -> \code{BlockwiseSparseReadIter}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{repr}}{...}

\item{\code{compress}}{...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BlockwiseSparseReadIter-new}{\code{BlockwiseSparseReadIter$new()}}
\item \href{#method-BlockwiseSparseReadIter-concat}{\code{BlockwiseSparseReadIter$concat()}}
\item \href{#method-BlockwiseSparseReadIter-clone}{\code{BlockwiseSparseReadIter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="reset"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-reset'><code>tiledbsoma::ReadIter$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="set_dim_points"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-set_dim_points'><code>tiledbsoma::ReadIter$set_dim_points()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="BlockwiseReadIterBase" data-id="read_complete"><a href='../../tiledbsoma/html/BlockwiseReadIterBase.html#method-BlockwiseReadIterBase-read_complete'><code>tiledbsoma::BlockwiseReadIterBase$read_complete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="BlockwiseReadIterBase" data-id="read_next"><a href='../../tiledbsoma/html/BlockwiseReadIterBase.html#method-BlockwiseReadIterBase-read_next'><code>tiledbsoma::BlockwiseReadIterBase$read_next()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockwiseSparseReadIter-new"></a>}}
\if{latex}{\out{\hypertarget{method-BlockwiseSparseReadIter-new}{}}}
\subsection{Method \code{new()}}{
...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockwiseSparseReadIter$new(
  sr,
  array,
  coords,
  axis,
  ...,
  repr = "T",
  compress = TRUE,
  reindex_disable_on_axis = NULL,
  eager = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sr}}{...}

\item{\code{array}}{...}

\item{\code{coords}}{...}

\item{\code{axis}}{...}

\item{\code{...}}{Ignored}

\item{\code{repr}}{...}

\item{\code{compress}}{...}

\item{\code{reindex_disable_on_axis}}{...}

\item{\code{eager}}{...}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockwiseSparseReadIter-concat"></a>}}
\if{latex}{\out{\hypertarget{method-BlockwiseSparseReadIter-concat}{}}}
\subsection{Method \code{concat()}}{
...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockwiseSparseReadIter$concat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockwiseSparseReadIter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlockwiseSparseReadIter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockwiseSparseReadIter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
