% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockwiseIter.R
\name{BlockwiseTableReadIter}
\alias{BlockwiseTableReadIter}
\title{SOMA Blockwise Read Iterator for Arrow Tables}
\description{
...
}
\keyword{internal}
\section{Super classes}{
\code{\link[tiledbsoma:ReadIter]{tiledbsoma::ReadIter}} -> \code{\link[tiledbsoma:BlockwiseReadIterBase]{tiledbsoma::BlockwiseReadIterBase}} -> \code{BlockwiseTableReadIter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BlockwiseTableReadIter-concat}{\code{BlockwiseTableReadIter$concat()}}
\item \href{#method-BlockwiseTableReadIter-clone}{\code{BlockwiseTableReadIter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="BlockwiseReadIterBase" data-id="initialize"><a href='../../tiledbsoma/html/BlockwiseReadIterBase.html#method-BlockwiseReadIterBase-initialize'><code>tiledbsoma::BlockwiseReadIterBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="BlockwiseReadIterBase" data-id="read_complete"><a href='../../tiledbsoma/html/BlockwiseReadIterBase.html#method-BlockwiseReadIterBase-read_complete'><code>tiledbsoma::BlockwiseReadIterBase$read_complete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="BlockwiseReadIterBase" data-id="read_next"><a href='../../tiledbsoma/html/BlockwiseReadIterBase.html#method-BlockwiseReadIterBase-read_next'><code>tiledbsoma::BlockwiseReadIterBase$read_next()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockwiseTableReadIter-concat"></a>}}
\if{latex}{\out{\hypertarget{method-BlockwiseTableReadIter-concat}{}}}
\subsection{Method \code{concat()}}{
...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockwiseTableReadIter$concat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockwiseTableReadIter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlockwiseTableReadIter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockwiseTableReadIter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
