% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordsStrider.R
\name{CoordsStrider}
\alias{CoordsStrider}
\alias{as.list.CoordsStrider}
\alias{length.CoordsStrider}
\alias{nextElem.CoordsStrider}
\alias{hasNext.CoordsStrider}
\title{The Coordinate Strider}
\usage{
\method{as.list}{CoordsStrider}(x, ...)

\method{length}{CoordsStrider}(x)

\method{nextElem}{CoordsStrider}(obj, ...)

\method{hasNext}{CoordsStrider}(obj, ...)
}
\description{
The \code{CoordsStrider} allows creating coordinate slices
in an iterated manner. Alternatively, it can chunk an existing vector of
coordinates
}
\note{
The \code{CoordsStrider} operates using
\link[bit64:integer64]{64-bit integer} objects; as such, accessing fields,
such as \code{strider$start} or \code{strider$stride} will return an
\code{integer64} object, which functions differently than a regular
\code{integer}. Use with caution and convert back to integers or numerics
as necessary
}
\examples{
(strider <- CoordsStrider$new(start = 1L, end = 200L, stride = 60L))
while (strider$has_next()) {
  str(strider$next_element())
}

(strider <- CoordsStrider$new(start = 1L, end = 200L, stride = 60L))
as.list(strider)

length(strider)

\dontshow{if (requireNamespace("itertools", quietly = TRUE)) withAutoprint(\{ # examplesIf}
(strider <- CoordsStrider$new(start = 1L, end = 200L, stride = 60L))
while (itertools::hasNext(strider)) {
  str(iterators::nextElem(strider))
}
\dontshow{\}) # examplesIf}
}
\keyword{internal}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{coords}}{If set, the coordinates to iterate over}

    \item{\code{full_coords}}{Every coordinate covered by the strider, without
advancing it}

    \item{\code{start}}{If set, the starting point of the iterated coordinates;
otherwise the minimum value of \code{self$coords}}

    \item{\code{end}}{If set, the end point of the iterated coordinates;
otherwise the maximum value of \code{self$coords}}

    \item{\code{stride}}{The stride, or how many coordinates to generate per
iteration; note: this field is settable, which will reset the iterator}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CoordsStrider-initialize}{\code{CoordsStrider$new()}}
    \item \href{#method-CoordsStrider-print}{\code{CoordsStrider$print()}}
    \item \href{#method-CoordsStrider-length}{\code{CoordsStrider$length()}}
    \item \href{#method-CoordsStrider-is_full_domain}{\code{CoordsStrider$is_full_domain()}}
    \item \href{#method-CoordsStrider-has_next}{\code{CoordsStrider$has_next()}}
    \item \href{#method-CoordsStrider-next_element}{\code{CoordsStrider$next_element()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-initialize}{}}}
\subsection{\code{CoordsStrider$new()}}{
  Create a coordinate strider
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoordsStrider$new(coords, ..., stride = NULL, start = NULL, end = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{coords}}{An integer vector of coordinates}
      \item{\code{...}}{Ignored}
      \item{\code{stride}}{The stride of how many coordinates to yield per iteration;
by default, will try to yield all coordinates per iteration}
      \item{\code{start}}{If \code{coords} is missing, the starting coordinate
to generate}
      \item{\code{end}}{If \code{coords} is missing, the ending coordinate
to generate}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-print"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-print}{}}}
\subsection{\code{CoordsStrider$print()}}{
  Print the coordinate strider to the screen
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoordsStrider$print()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-length"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-length}{}}}
\subsection{\code{CoordsStrider$length()}}{
  Get the length (span) of the coordinates within the strider
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoordsStrider$length()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The length (span) of the coordinate strider
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-is_full_domain"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-is_full_domain}{}}}
\subsection{\code{CoordsStrider$is_full_domain()}}{
  Determine if this strider covers the full domain
\code{[0, extent)} of a dimension
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoordsStrider$is_full_domain(extent)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{extent}}{The extent (shape) of the dimension}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{TRUE} if the strider is a range strider spanning
\code{0:(extent - 1)}, otherwise \code{FALSE}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-has_next"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-has_next}{}}}
\subsection{\code{CoordsStrider$has_next()}}{
  Determine if there are more coordinates to yield
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoordsStrider$has_next()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{TRUE} if there are more coordinates to yield or
\code{FALSE} if otherwise
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-next_element"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-next_element}{}}}
\subsection{\code{CoordsStrider$next_element()}}{
  Generate the next set of coordinates to yield. If there are
no more coordinates to yield, raises a \code{stopIteration} error
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoordsStrider$next_element()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    An integer vector of the next set of coordinates
  }
}

}
