% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset_seurat_pbmc3k}
\alias{dataset_seurat_pbmc3k}
\title{Seurat 3k PBMCs from 10x Genomics}
\usage{
dataset_seurat_pbmc3k()
}
\value{
a \code{\link[SeuratObject:Seurat-class]{SeuratObject::Seurat}} object
}
\description{
Create a \code{\link[SeuratObject:Seurat-class]{SeuratObject::Seurat}} object containing the widely used 3k PBMCs
dataset from 10x Genomics.
}
\seealso{
https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k
}
