% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{delete_metadata}
\alias{delete_metadata}
\title{Delete metadata for given key}
\usage{
delete_metadata(uri, key, ctxxp)
}
\arguments{
\item{uri}{The array URI}

\item{key}{The array metadata key}

\item{ctxxp}{An external pointer to the SOMAContext wrapper}
}
\description{
Delete metadata for given key
}
