% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseReadIter.R
\name{DenseReadIter}
\alias{DenseReadIter}
\title{DenseReadIter}
\description{
\code{DenseReadIter} is a class that allows for iteration over
a reads on \link{SOMADenseNDArray}.
Iteration chunks are retrieved as 0-based Views of \link[base]{matrix}.
}
\section{Super class}{
\code{\link[tiledbsoma:ReadIter]{tiledbsoma::ReadIter}} -> \code{DenseReadIter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DenseReadIter-new}{\code{DenseReadIter$new()}}
\item \href{#method-DenseReadIter-concat}{\code{DenseReadIter$concat()}}
\item \href{#method-DenseReadIter-clone}{\code{DenseReadIter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="read_complete"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-read_complete'><code>tiledbsoma::ReadIter$read_complete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="read_next"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-read_next'><code>tiledbsoma::ReadIter$read_next()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DenseReadIter-new"></a>}}
\if{latex}{\out{\hypertarget{method-DenseReadIter-new}{}}}
\subsection{Method \code{new()}}{
Create (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DenseReadIter$new(
  uri,
  config,
  dim_points = NULL,
  result_order = NULL,
  loglevel = "auto"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{Character value with URI path to a SOMADataFrame or SOMASparseNDArray}

\item{\code{config}}{character vector containing TileDB config.
no query condition}

\item{\code{dim_points}}{Optional named list with vector of data points to select on the given
dimension(s). Each dimension can be one entry in the list.}

\item{\code{loglevel}}{Character value with the desired logging level, defaults to \sQuote{auto}}

\item{\code{repr}}{Optional one-character code for sparse matrix representation type
which lets prior setting prevail, any other value is set as new logging level.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DenseReadIter-concat"></a>}}
\if{latex}{\out{\hypertarget{method-DenseReadIter-concat}{}}}
\subsection{Method \code{concat()}}{
Concatenate remainder of iterator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DenseReadIter$concat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{matrixZeroBasedView}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DenseReadIter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DenseReadIter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DenseReadIter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
