% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{dim.matrixZeroBasedView}
\alias{dim.matrixZeroBasedView}
\title{dim}
\usage{
\method{dim}{matrixZeroBasedView}(x)
}
\arguments{
\item{x}{The zero-based matrix view.}
}
\value{
The dimensions of the matrix.
}
\description{
dim
}
