% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephemeral.R
\name{EphemeralCollectionBase}
\alias{EphemeralCollectionBase}
\title{Ephemeral Collection Base}
\description{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_cls("collection", TRUE)}
}
\keyword{internal}
\section{Super classes}{
\code{\link[tiledbsoma:TileDBObject]{tiledbsoma::TileDBObject}} -> \code{\link[tiledbsoma:TileDBGroup]{tiledbsoma::TileDBGroup}} -> \code{\link[tiledbsoma:SOMACollectionBase]{tiledbsoma::SOMACollectionBase}} -> \code{EphemeralCollectionBase}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{uri}}{\dQuote{\code{ephemeral-collection:<MEMORY_ADDRESS>}}}

\item{\code{soma_type}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_field()}}

\item{\code{platform_config}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_field()}}

\item{\code{tiledbsoma_ctx}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_field()}}

\item{\code{object}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_field()}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EphemeralCollectionBase-new}{\code{EphemeralCollectionBase$new()}}
\item \href{#method-EphemeralCollectionBase-create}{\code{EphemeralCollectionBase$create()}}
\item \href{#method-EphemeralCollectionBase-open}{\code{EphemeralCollectionBase$open()}}
\item \href{#method-EphemeralCollectionBase-close}{\code{EphemeralCollectionBase$close()}}
\item \href{#method-EphemeralCollectionBase-exists}{\code{EphemeralCollectionBase$exists()}}
\item \href{#method-EphemeralCollectionBase-print}{\code{EphemeralCollectionBase$print()}}
\item \href{#method-EphemeralCollectionBase-get_tiledb_config}{\code{EphemeralCollectionBase$get_tiledb_config()}}
\item \href{#method-EphemeralCollectionBase-length}{\code{EphemeralCollectionBase$length()}}
\item \href{#method-EphemeralCollectionBase-names}{\code{EphemeralCollectionBase$names()}}
\item \href{#method-EphemeralCollectionBase-set}{\code{EphemeralCollectionBase$set()}}
\item \href{#method-EphemeralCollectionBase-get}{\code{EphemeralCollectionBase$get()}}
\item \href{#method-EphemeralCollectionBase-remove}{\code{EphemeralCollectionBase$remove()}}
\item \href{#method-EphemeralCollectionBase-set_metadata}{\code{EphemeralCollectionBase$set_metadata()}}
\item \href{#method-EphemeralCollectionBase-get_metadata}{\code{EphemeralCollectionBase$get_metadata()}}
\item \href{#method-EphemeralCollectionBase-add_new_collection}{\code{EphemeralCollectionBase$add_new_collection()}}
\item \href{#method-EphemeralCollectionBase-add_new_dataframe}{\code{EphemeralCollectionBase$add_new_dataframe()}}
\item \href{#method-EphemeralCollectionBase-add_new_dense_ndarray}{\code{EphemeralCollectionBase$add_new_dense_ndarray()}}
\item \href{#method-EphemeralCollectionBase-add_new_sparse_ndarray}{\code{EphemeralCollectionBase$add_new_sparse_ndarray()}}
\item \href{#method-EphemeralCollectionBase-clone}{\code{EphemeralCollectionBase$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="class"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-class'><code>tiledbsoma::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="is_open"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-is_open'><code>tiledbsoma::TileDBObject$is_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="mode"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-mode'><code>tiledbsoma::TileDBObject$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="to_data_frame"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-to_data_frame'><code>tiledbsoma::TileDBGroup$to_data_frame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="to_list"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-to_list'><code>tiledbsoma::TileDBGroup$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-new}{}}}
\subsection{Method \code{new()}}{
Create an ephemeral collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-create"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-create}{}}}
\subsection{Method \code{create()}}{
Create a new, empty ephemeral collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$create()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a new ephemeral collection of class \code{class(self)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-open"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-open}{}}}
\subsection{Method \code{open()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$open(mode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_param()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_error()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-close"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-close}{}}}
\subsection{Method \code{close()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisibly returns \code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-exists"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-exists}{}}}
\subsection{Method \code{exists()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$exists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns \code{FALSE} as ephemeral collections do not
exist on-disk
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-print"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-print}{}}}
\subsection{Method \code{print()}}{
Special method for printing object representation to console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Prints details about the ephemeral collection and invisibly
returns itself
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-get_tiledb_config"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-get_tiledb_config}{}}}
\subsection{Method \code{get_tiledb_config()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$get_tiledb_config(param = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_param()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns \code{NULL} as ephemeral collections do not have an
on-disk configuration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-length"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-length}{}}}
\subsection{Method \code{length()}}{
Retrieve the number of items in the collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$length()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The length of the collection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-names"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-names}{}}}
\subsection{Method \code{names()}}{
Retrieve the names of members. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} vector of member names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-set"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-set}{}}}
\subsection{Method \code{set()}}{
Add object to an ephemeral collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$set(object, name = NULL, relative = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A TileDB object (eg. \code{\link{TileDBGroup}}) to add
to the collection}

\item{\code{name}}{A name to add \code{object} as}

\item{\code{relative}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_param()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[chainable] Invisibly returns \code{self} with \code{object}
added as \code{name}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-get"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-get}{}}}
\subsection{Method \code{get()}}{
Get objects from an ephemeral collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$get(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of object in the collection to get}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object named \code{name}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-remove"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-remove}{}}}
\subsection{Method \code{remove()}}{
Remove objects from an ephemeral collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$remove(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of object to remove from the collection}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[chainable] Invisibly returns \code{self} with the object at
\code{name} removed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-set_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-set_metadata}{}}}
\subsection{Method \code{set_metadata()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$set_metadata(metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_param()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_error()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$get_metadata(key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_param()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An empty list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-add_new_collection"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-add_new_collection}{}}}
\subsection{Method \code{add_new_collection()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$add_new_collection(object, key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object, key}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_param()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_error()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-add_new_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-add_new_dataframe}{}}}
\subsection{Method \code{add_new_dataframe()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$add_new_dataframe(key, schema, index_column_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key, schema, index_column_names}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_param()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_error()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-add_new_dense_ndarray"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-add_new_dense_ndarray}{}}}
\subsection{Method \code{add_new_dense_ndarray()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$add_new_dense_ndarray(key, type, shape)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key, type, shape}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_param()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_error()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-add_new_sparse_ndarray"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-add_new_sparse_ndarray}{}}}
\subsection{Method \code{add_new_sparse_ndarray()}}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_desc()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$add_new_sparse_ndarray(key, type, shape)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key, type, shape}}{\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_param()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_error()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralCollectionBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralCollectionBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralCollectionBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
