% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephemeral.R
\name{EphemeralExperiment}
\alias{EphemeralExperiment}
\title{Ephemeral SOMA Experiment}
\description{
\Sexpr[results=rd]{tiledbsoma:::rd_ephemeral_cls("experiment")}
}
\keyword{internal}
\section{Super classes}{
\code{tiledbsoma::TileDBObject} -> \code{tiledbsoma::TileDBGroup} -> \code{tiledbsoma::SOMACollectionBase} -> \code{tiledbsoma::EphemeralCollectionBase} -> \code{EphemeralExperiment}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{obs}}{\Sexpr[results=rd]{tiledbsoma:::rd_soma_field("obs")}}

\item{\code{ms}}{\Sexpr[results=rd]{tiledbsoma:::rd_soma_field("ms")}}

\item{\code{soma_type}}{The SOMA object type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EphemeralExperiment-clone}{\code{EphemeralExperiment$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="class"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-class'><code>tiledbsoma::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="is_open"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-is_open'><code>tiledbsoma::TileDBObject$is_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="mode"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-mode'><code>tiledbsoma::TileDBObject$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="to_data_frame"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-to_data_frame'><code>tiledbsoma::TileDBGroup$to_data_frame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="to_list"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-to_list'><code>tiledbsoma::TileDBGroup$to_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="add_new_collection"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-add_new_collection'><code>tiledbsoma::EphemeralCollectionBase$add_new_collection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="add_new_dataframe"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-add_new_dataframe'><code>tiledbsoma::EphemeralCollectionBase$add_new_dataframe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="add_new_dense_ndarray"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-add_new_dense_ndarray'><code>tiledbsoma::EphemeralCollectionBase$add_new_dense_ndarray()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="add_new_sparse_ndarray"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-add_new_sparse_ndarray'><code>tiledbsoma::EphemeralCollectionBase$add_new_sparse_ndarray()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="close"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-close'><code>tiledbsoma::EphemeralCollectionBase$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="create"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-create'><code>tiledbsoma::EphemeralCollectionBase$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="exists"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-exists'><code>tiledbsoma::EphemeralCollectionBase$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="get"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-get'><code>tiledbsoma::EphemeralCollectionBase$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="get_metadata"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-get_metadata'><code>tiledbsoma::EphemeralCollectionBase$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="get_tiledb_config"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-get_tiledb_config'><code>tiledbsoma::EphemeralCollectionBase$get_tiledb_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="initialize"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-initialize'><code>tiledbsoma::EphemeralCollectionBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="length"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-length'><code>tiledbsoma::EphemeralCollectionBase$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="names"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-names'><code>tiledbsoma::EphemeralCollectionBase$names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="open"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-open'><code>tiledbsoma::EphemeralCollectionBase$open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="print"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-print'><code>tiledbsoma::EphemeralCollectionBase$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="remove"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-remove'><code>tiledbsoma::EphemeralCollectionBase$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="set"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-set'><code>tiledbsoma::EphemeralCollectionBase$set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="EphemeralCollectionBase" data-id="set_metadata"><a href='../../tiledbsoma/html/EphemeralCollectionBase.html#method-EphemeralCollectionBase-set_metadata'><code>tiledbsoma::EphemeralCollectionBase$set_metadata()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EphemeralExperiment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EphemeralExperiment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EphemeralExperiment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
