% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{example-datasets}
\alias{example-datasets}
\alias{list_datasets}
\alias{extract_dataset}
\alias{load_dataset}
\title{SOMA Example Datasets}
\usage{
list_datasets()

extract_dataset(name, dir = tempdir())

load_dataset(name, dir = tempdir(), tiledbsoma_ctx = NULL, context = NULL)
}
\arguments{
\item{name}{The name of the dataset.}

\item{dir}{The directory where the dataset will be extracted to
(default: \code{\link[base]{tempdir}()}).}

\item{tiledbsoma_ctx}{Optional (DEPRECATED) TileDB \dQuote{Context} object
that defaults to \code{NULL}.}

\item{context}{Optional \code{SOMAContext} object used for TileDB operations.
If a context is not provided, then the default context will be used. Call
\code{set_default_context} once before other SOMA operations to configure
the default context.}
}
\value{
\code{list_datasets()}: returns a character vector of the
available datasets.

\code{extract_dataset()}: returns the path to the extracted dataset.

\code{load_dataset()}: returns a SOMA object.
}
\description{
Access example SOMA objects bundled with the tiledbsoma package.\cr
\cr
Use \code{list_datasets()} to list the available datasets and
\code{load_dataset()} to load a dataset into memory using the appropriate
SOMA class. The \code{extract_dataset()} function returns the path to the
extracted dataset without loading it into memory.
}
\details{
The SOMA objects are stored as \code{tar.gz} files in the package's
\code{extdata} directory. Calling \code{load_dataset()} extracts the
\code{tar.gz} file to the specified \code{dir}, inspects its metadata to
determine the appropriate SOMA class to instantiate, and returns the
SOMA object.
}
\examples{
list_datasets()

\dontshow{if (requireNamespace("withr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
dir <- withr::local_tempfile(pattern = "pbmc-small")
dir.create(dir, recursive = TRUE)
dest <- extract_dataset("soma-exp-pbmc-small", dir)
list.files(dest)
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("withr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
dir <- withr::local_tempfile(pattern = "pbmc_small")
dir.create(dir, recursive = TRUE)
(exp <- load_dataset("soma-exp-pbmc-small", dir))

\dontshow{
exp$close()
}
\dontshow{\}) # examplesIf}
}
