% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_all_metadata}
\alias{get_all_metadata}
\title{Read all metadata (as named character vector)}
\usage{
get_all_metadata(uri, ctxxp)
}
\arguments{
\item{uri}{The array URI}

\item{ctxxp}{An external pointer to the SOMAContext wrapper}
}
\description{
This function assumes that all metadata is in fact stored as strings. It will error
if a different datatype is encountered.
}
