% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAContext.R
\name{get_default_context}
\alias{get_default_context}
\title{Get the Default SOMA Context}
\usage{
get_default_context()
}
\value{
The context that will be used for TileDB-SOMA API when no context is provided by the user.
}
\description{
Retrieve the current default \code{\link{SOMAContext}} used by TileDB-SOMA
operations.
}
\details{
This function returns the context that was either:
\itemize{
\item Explicitly set via \code{\link{set_default_context}}, or
\item Automatically created when a SOMA object was first created
}

An error is raised if no default context is set.
}
