% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_metadata}
\alias{get_metadata}
\title{Read metadata (as a string)}
\usage{
get_metadata(uri, key, is_array, ctxxp)
}
\arguments{
\item{uri}{The array URI}

\item{key}{The array metadata key}

\item{is_array}{A boolean to indicate array or group}

\item{ctxxp}{An external pointer to the SOMAContext wrapper}
}
\description{
Read metadata (as a string)
}
