% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/getData.R
\name{get_column_names}
\alias{get_column_names}
\alias{export_column_direct}
\alias{get_table}
\alias{get_column}
\title{Access a DataFrame or Single Column From a Given URI}
\usage{
get_column_names(uri)

export_column_direct(uri, colnames)

get_table(uri)

get_column(uri, column)
}
\arguments{
\item{uri}{Character value with URI path to a SOMA data set}

\item{colnames}{Vector of character value with the name of the columns to retrieve}

\item{column}{Character value with the name of the column to retrieve}
}
\value{
The selected data frame (or column) from the given data set
}
\description{
These functions access a given SOMA URI and extract, respectively, the
complete data.frame or named column from the data set.
}
\examples{
\dontrun{
uri <- "test/soco/pbmc3k_processed/obs"
column <-  "n_counts"
summary(get_table(uri))
}
}
