% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getColumn}
\alias{getColumn}
\alias{getTable}
\title{Access a DataFrame or Single Column From a Given URI}
\usage{
getColumn(uri, column)

getTable(uri = "test/soco/pbmc3k_processed/obs")
}
\arguments{
\item{uri}{Character value with URI path to a SOMA data set}

\item{column}{Character value with the name of the column to retrieve}
}
\value{
The selected column from the give data set
}
\description{
These functions access a given SOMA URI and extract, respectively, the
complete data.frame or named column from the data set.
}
\examples{
\dontrun{
uri <- "test/soco/pbmc3k_processed/obs"
column <-  "n_counts"
summary(getColumn(uri, column)
summary(getTable(uri))
}
}
