% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{libtiledbsoma_version}
\alias{libtiledbsoma_version}
\title{libtiledbsoma version information}
\usage{
libtiledbsoma_version()
}
\description{
Returns a string with version information for libtiledbsoma and the linked TileDB Embedded library.
}
