% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingBase.R
\name{MappingBase}
\alias{MappingBase}
\title{R6 Base Mapping Type}
\description{
R6 Base Mapping Type

R6 Base Mapping Type
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MappingBase-new}{\code{MappingBase$new()}}
\item \href{#method-MappingBase-keys}{\code{MappingBase$keys()}}
\item \href{#method-MappingBase-values}{\code{MappingBase$values()}}
\item \href{#method-MappingBase-items}{\code{MappingBase$items()}}
\item \href{#method-MappingBase-get}{\code{MappingBase$get()}}
\item \href{#method-MappingBase-set}{\code{MappingBase$set()}}
\item \href{#method-MappingBase-setv}{\code{MappingBase$setv()}}
\item \href{#method-MappingBase-clone}{\code{MappingBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappingBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-MappingBase-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingBase$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappingBase-keys"></a>}}
\if{latex}{\out{\hypertarget{method-MappingBase-keys}{}}}
\subsection{Method \code{keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingBase$keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The keys of the map
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappingBase-values"></a>}}
\if{latex}{\out{\hypertarget{method-MappingBase-values}{}}}
\subsection{Method \code{values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingBase$values()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The values of the map
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappingBase-items"></a>}}
\if{latex}{\out{\hypertarget{method-MappingBase-items}{}}}
\subsection{Method \code{items()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingBase$items()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappingBase-get"></a>}}
\if{latex}{\out{\hypertarget{method-MappingBase-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingBase$get(key, default = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{...}

\item{\code{default}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappingBase-set"></a>}}
\if{latex}{\out{\hypertarget{method-MappingBase-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingBase$set(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{...}

\item{\code{value}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[chainable] Invisibly returns \code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappingBase-setv"></a>}}
\if{latex}{\out{\hypertarget{method-MappingBase-setv}{}}}
\subsection{Method \code{setv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingBase$setv(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[chainable] Invisibly returns \code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappingBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MappingBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
