% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{matrixZeroBasedView}
\alias{matrixZeroBasedView}
\title{Zero-based matrix shim}
\usage{
matrixZeroBasedView(one_based_matrix)
}
\arguments{
\item{one_based_matrix}{matrix or Matrix::sparseMatrix}
}
\value{
Shim providing elemental access to the matrix using zero-based indexes.
}
\description{
Zero-based matrix shim
}
