% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{ncol.matrixZeroBasedView}
\alias{ncol.matrixZeroBasedView}
\title{ncol}
\usage{
ncol.matrixZeroBasedView(x)
}
\arguments{
\item{x}{The zero-based matrix view.}
}
\value{
Matrix column count.
}
\description{
ncol
}
