% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{nrow.matrixZeroBasedView}
\alias{nrow.matrixZeroBasedView}
\title{nrow}
\usage{
nrow.matrixZeroBasedView(x)
}
\arguments{
\item{x}{The zero-based matrix view.}
}
\value{
Matrix row count.
}
\description{
nrow
}
