% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{Ops.matrixZeroBasedView}
\alias{Ops.matrixZeroBasedView}
\title{\itemize{
\item 
}}
\usage{
\method{Ops}{matrixZeroBasedView}(e1, e2 = NULL)
}
\arguments{
\item{e1}{left side}

\item{e2}{right sidet}
}
\value{
results of sum
}
\description{
\itemize{
\item 
}
}
