% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_matrix.R
\name{pad_matrix}
\alias{pad_matrix}
\alias{pad_matrix.matrix}
\title{Pad a Matrix}
\usage{
pad_matrix(x, ...)

\method{pad_matrix}{matrix}(x, rowidx, colidx, shape, sparse = FALSE, ...)
}
\arguments{
\item{x}{A matrix-like object}

\item{...}{Arguments passed to other methods}

\item{rowidx, colidx}{Integer indices for rows and columns from
\code{x} to \code{shape}}

\item{shape}{Target shape to pad out to}

\item{sparse}{Return a \link[Matrix:TsparseMatrix-class]{sparse matrix}}
}
\value{
A matrix containing the data of \code{x} padded out to \code{shape}
}
\description{
Pad a Matrix
}
\keyword{internal}
