% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlatformConfig.R
\name{PlatformConfig}
\alias{PlatformConfig}
\title{Platform Configuration}
\description{
Platform Configuration

Platform Configuration
}
\section{Super class}{
\code{tiledbsoma::MappingBase} -> \code{PlatformConfig}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlatformConfig-platforms}{\code{PlatformConfig$platforms()}}
\item \href{#method-PlatformConfig-ops}{\code{PlatformConfig$ops()}}
\item \href{#method-PlatformConfig-get}{\code{PlatformConfig$get()}}
\item \href{#method-PlatformConfig-get_ops}{\code{PlatformConfig$get_ops()}}
\item \href{#method-PlatformConfig-set}{\code{PlatformConfig$set()}}
\item \href{#method-PlatformConfig-clone}{\code{PlatformConfig$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="initialize"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-initialize'><code>tiledbsoma::MappingBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="items"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-items'><code>tiledbsoma::MappingBase$items()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="keys"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-keys'><code>tiledbsoma::MappingBase$keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="length"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-length'><code>tiledbsoma::MappingBase$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="print"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-print'><code>tiledbsoma::MappingBase$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="setv"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-setv'><code>tiledbsoma::MappingBase$setv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="to_list"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-to_list'><code>tiledbsoma::MappingBase$to_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="update"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-update'><code>tiledbsoma::MappingBase$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="values"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-values'><code>tiledbsoma::MappingBase$values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlatformConfig-platforms"></a>}}
\if{latex}{\out{\hypertarget{method-PlatformConfig-platforms}{}}}
\subsection{Method \code{platforms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlatformConfig$platforms()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlatformConfig-ops"></a>}}
\if{latex}{\out{\hypertarget{method-PlatformConfig-ops}{}}}
\subsection{Method \code{ops()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlatformConfig$ops(platform = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{platform}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlatformConfig-get"></a>}}
\if{latex}{\out{\hypertarget{method-PlatformConfig-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlatformConfig$get(platform, op = NULL, key = NULL, default = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{platform}}{...}

\item{\code{op}}{...}

\item{\code{key}}{...}

\item{\code{default}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlatformConfig-get_ops"></a>}}
\if{latex}{\out{\hypertarget{method-PlatformConfig-get_ops}{}}}
\subsection{Method \code{get_ops()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlatformConfig$get_ops(platform)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{platform}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlatformConfig-set"></a>}}
\if{latex}{\out{\hypertarget{method-PlatformConfig-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlatformConfig$set(platform, op, key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{platform}}{...}

\item{\code{op}}{...}

\item{\code{key}}{...}

\item{\code{value}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlatformConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlatformConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlatformConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
