% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{print.matrixZeroBasedView}
\alias{print.matrixZeroBasedView}
\title{print}
\usage{
\method{print}{matrixZeroBasedView}(x)
}
\arguments{
\item{x}{The zero-based matrix view.}
}
\value{
Matrix column count.
}
\description{
print
}
