% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-arrow.R
\name{r_type_from_arrow_type}
\alias{r_type_from_arrow_type}
\alias{r_type_from_arrow_type.Schema}
\alias{r_type_from_arrow_type.Field}
\alias{r_type_from_arrow_type.DataType}
\title{Get the \R Type from an Arrow Type}
\usage{
r_type_from_arrow_type(x)

\method{r_type_from_arrow_type}{Schema}(x)

\method{r_type_from_arrow_type}{Field}(x)

\method{r_type_from_arrow_type}{DataType}(x)
}
\arguments{
\item{x}{An \CRANpkg{Arrow} \link[arrow:Schema]{schema},
\link[arrow:Field]{field}, or \link[arrow:infer_type]{data type}}
}
\value{
If \code{x} is a \link[arrow:infer_type]{data type}, a single
character value giving the \R \link[base:typeof]{type} of \code{x}; if no
corresponding \R type, returns the \CRANpkg{Arrow} type name

If \code{x} is a \link[arrow:Field]{field}, a single named character
vector with the name being the field name and the value being the \R
\link[base:typeof]{type}

If \code{x} is a \link[arrow:Schema]{schema}, a named vector where
the names are field names and the values are the \R \link[base:typeof]{types}
of each field
}
\description{
Get the \R Type from an Arrow Type
}
\keyword{internal}
