% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordsStrider.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{nextElem}
\alias{hasNext}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{iterators}{\code{\link[iterators]{nextElem}}}

  \item{itertools}{\code{\link[itertools]{hasNext}}}
}}

