% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reindexer.R
\name{Reindexer}
\alias{Reindexer}
\title{Reindexer Class}
\description{
Reindexer Class

Reindexer Class
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Reindexer-new}{\code{Reindexer$new()}}
\item \href{#method-Reindexer-lookup}{\code{Reindexer$lookup()}}
\item \href{#method-Reindexer-clone}{\code{Reindexer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reindexer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Reindexer-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Reindexer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reindexer$new(map_keys, nb_threads = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map_keys}}{An integer64 vector with keys to be mapped by the class}

\item{\code{nb_threads}}{The number of threads to use, defaults to four}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reindexer-lookup"></a>}}
\if{latex}{\out{\hypertarget{method-Reindexer-lookup}{}}}
\subsection{Method \code{lookup()}}{
Lookup a vector of keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reindexer$lookup(lookup_keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookup_keys}}{An integer64 vector with values to be looked up}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of the same length with index positions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reindexer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Reindexer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reindexer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
