% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAContext.R
\name{set_default_context}
\alias{set_default_context}
\title{Create and return a global default context object}
\usage{
set_default_context(config = NULL, replace = FALSE)
}
\arguments{
\item{config}{...}

\item{replace}{Allow replacing an existing default context with a new default context}
}
\value{
Invisibly, the default default context object.
}
\description{
If the global context was already set, an error will be raised unless
\code{replace=True}. Setting a new default context will not change the
context for TileDB-SOMA objects that were already created.
}
