% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_log_level}
\alias{set_log_level}
\title{Set TileDB-SOMA Logging Level}
\usage{
set_log_level(level)
}
\arguments{
\item{level}{A character value with logging level understood by
\sQuote{spdlog} such as \dQuote{trace}, \dQuote{debug}, \dQuote{info}, or
\dQuote{warn}}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Set the logging level for the \R package and underlying C++ library
}
