% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_metadata}
\alias{set_metadata}
\title{Set metadata (as a string)}
\usage{
set_metadata(uri, key, value, ctxxp)
}
\arguments{
\item{uri}{The array URI}

\item{key}{The array metadata key}

\item{value}{The metadata value}

\item{ctxxp}{An external pointer to the SOMAContext wrapper}
}
\description{
Set metadata (as a string)
}
