% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tiledb.R
\name{show_package_versions}
\alias{show_package_versions}
\title{Display Package Versions}
\usage{
show_package_versions()
}
\description{
This helperfunction prints package information suitable for assisting with bug reports.
}
