% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soma_array_reader}
\alias{soma_array_reader}
\alias{nnz}
\alias{shape}
\title{Read SOMA Data From a Given URI}
\usage{
soma_array_reader(
  uri,
  colnames = NULL,
  qc = NULL,
  dim_points = NULL,
  dim_ranges = NULL,
  batch_size = "auto",
  result_order = "auto",
  loglevel = "auto",
  config = NULL
)

nnz(uri, config = NULL)

shape(uri, config = NULL)
}
\arguments{
\item{uri}{Character value with URI path to a SOMA data set}

\item{colnames}{Optional vector of character value with the name of the columns to retrieve}

\item{qc}{Optional external Pointer object to TileDB Query Condition, defaults to \sQuote{NULL} i.e.
no query condition}

\item{dim_points}{Optional named list with vector of data points to select on the given
dimension(s). Each dimension can be one entry in the list.}

\item{dim_ranges}{Optional named list with two-column matrix where each row select a range
for the given dimension. Each dimension can be one entry in the list.}

\item{batch_size}{Character value with the desired batch size, defaults to \sQuote{auto}}

\item{result_order}{Character value with the desired result order, defaults to \sQuote{auto}}

\item{loglevel}{Character value with the desired logging level, defaults to \sQuote{auto}
which lets prior setting prevail, any other value is set as new logging level.}

\item{config}{Optional character vector containing TileDB config.}
}
\value{
A List object with two pointers to Arrow array data and schema is returned
}
\description{
This functions access a given SOMA URI and returns a complete data.frame. It does
not iterate; if your data is large than the initial read size consider the \code{sr_*}
functions.
}
\examples{
\dontrun{
uri <- extract_dataset("soma-dataframe-pbmc3k-processed-obs")
z <- soma_array_reader(uri)
arrow::RecordBatch$import_from_c(z$array_data, z$schema)
}
}
