% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-readerTransformers.R
\name{soma_array_to_arrow_table}
\alias{soma_array_to_arrow_table}
\title{Transformer function: SOMAArray to Arrow table}
\usage{
soma_array_to_arrow_table(x)
}
\arguments{
\item{x}{A List object with two pointers to Arrow array data and schema}
}
\value{
arrow::\link[arrow]{Table}
}
\description{
Converts the results of a \link{soma_array_reader} or \link{sr_next} to
an arrow::\link[arrow]{Table}
}
