% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soma_reader}
\alias{soma_reader}
\alias{nnz}
\title{Read SOMA Data From a Given URI}
\usage{
soma_reader(
  uri,
  colnames = NULL,
  qc = NULL,
  dim_points = NULL,
  dim_ranges = NULL,
  loglevel = "warn"
)

nnz(uri)
}
\arguments{
\item{uri}{Character value with URI path to a SOMA data set}

\item{colnames}{Optional vector of character value with the name of the columns to retrieve}

\item{qc}{Optional external Pointer object to TileDB Query Condition, defaults to \sQuote{NULL} i.e.
no query condition}

\item{dim_points}{Optional named list with vector of data points to select on the given
dimension(s). Each dimension can be one entry in the list.}

\item{dim_ranges}{Optional named list with two-column matrix where each row select a range
for the given dimension. Each dimension can be one entry in the list.}

\item{loglevel}{Character value with the desired logging level, defaults to \sQuote{warn}}
}
\value{
An Arrow data structure is returned
}
\description{
This functions access a given SOMA URI and returns a complete data.frame. It does
not iterate; if your data is large than the initial read size consider the \code{sr_*}
functions.
}
\examples{
\dontrun{
uri <- "test/soco/pbmc3k_processed/obs"
z <- soma_reader(uri)
tb <- arrow::as_arrow_table(arch::from_arch_array(z, arrow::RecordBatch))
}
}
