% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAAxisIndexer.R
\name{SOMAAxisIndexer}
\alias{SOMAAxisIndexer}
\title{\code{SOMAExperiment} Axis Query Result Indexer}
\description{
Index \code{obs}/\code{var} soma_joinids for a given query result.

Retrieve the index of the given \code{obs} or \code{var} coordinates in the query
result. Coordinates outside of the query result will return
\code{\link[arrow:data-type]{arrow::null()}}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMAAxisIndexer-new}{\code{SOMAAxisIndexer$new()}}
\item \href{#method-SOMAAxisIndexer-by_obs}{\code{SOMAAxisIndexer$by_obs()}}
\item \href{#method-SOMAAxisIndexer-by_var}{\code{SOMAAxisIndexer$by_var()}}
\item \href{#method-SOMAAxisIndexer-clone}{\code{SOMAAxisIndexer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAAxisIndexer-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAAxisIndexer-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SOMAAxisIndexer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAAxisIndexer$new(query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{The \code{\link{SOMAExperimentAxisQuery}} object to build indices for.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAAxisIndexer-by_obs"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAAxisIndexer-by_obs}{}}}
\subsection{Method \code{by_obs()}}{
Get the index of the given \code{obs} coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAAxisIndexer$by_obs(coords)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coords}}{vector or \code{\link[arrow:array-class]{arrow::Array}} of numeric coordinates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAAxisIndexer-by_var"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAAxisIndexer-by_var}{}}}
\subsection{Method \code{by_var()}}{
Get the index of the given \code{var} coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAAxisIndexer$by_var(coords)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coords}}{vector or \code{\link[arrow:array-class]{arrow::Array}} of numeric coordinates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAAxisIndexer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAAxisIndexer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAAxisIndexer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
