% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAAxisQuery.R
\name{SOMAAxisQuery}
\alias{SOMAAxisQuery}
\title{SOMA Axis Query}
\description{
Construct a single-axis query object with a combination of
coordinates and/or value filters for use with \code{\link{SOMAExperimentAxisQuery}}.
(lifecycle: maturing)

Per dimension, the \code{SOMAAxisQuery} can have value of:
\itemize{
\item None (i.e., \code{coords = NULL} and \code{value_filter = NULL}) - read all values
\item Coordinates - a set of coordinates on the axis dataframe index, expressed
in any type or format supported by \code{\link{SOMADataFrame}}'s \code{read()} method.
\item A SOMA \code{value_filter} across columns in the axis dataframe, expressed as
string
\item Or, a combination of coordinates and value filter.
}
}
\seealso{
\code{\link[tiledb:parse_query_condition]{tiledb::parse_query_condition()}} for more information about valid
value filters.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coords}}{The coordinates for the query.}

\item{\code{value_filter}}{The value filter for the query.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMAAxisQuery-new}{\code{SOMAAxisQuery$new()}}
\item \href{#method-SOMAAxisQuery-clone}{\code{SOMAAxisQuery$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAAxisQuery-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAAxisQuery-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SOMAAxisQuery} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAAxisQuery$new(value_filter = NULL, coords = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value_filter}}{Optional string containing a logical expression that
is used to filter the returned values.}

\item{\code{coords}}{Optional indices specifying the rows to read: either a
vector of the appropriate type or a named list of vectors corresponding
to each dimension.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAAxisQuery-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAAxisQuery-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAAxisQuery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
