% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAAxisQueryResult.R
\name{SOMAAxisQueryResult}
\alias{SOMAAxisQueryResult}
\title{\code{SOMAExperiment} Axis Query Result}
\description{
Access \code{\link{SOMAExperimentAxisQuery}} results.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{obs}}{\code{\link[arrow:Table-class]{arrow::Table}} containing \code{obs} query slice.}

\item{\code{var}}{\code{\link[arrow:Table-class]{arrow::Table}} containing \code{var} query slice.
\code{measurement_name}.}

\item{\code{X_layers}}{named list of \code{\link[arrow:Table-class]{arrow::Table}}s for each \code{X} layer.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMAAxisQueryResult-new}{\code{SOMAAxisQueryResult$new()}}
\item \href{#method-SOMAAxisQueryResult-clone}{\code{SOMAAxisQueryResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAAxisQueryResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAAxisQueryResult-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SOMAAxisQueryResult} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAAxisQueryResult$new(obs, var, X_layers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obs, var}}{\code{\link[arrow:Table-class]{arrow::Table}} containing \code{obs} or \code{var} query slice.}

\item{\code{X_layers}}{named list of \code{\link[arrow:Table-class]{arrow::Table}}s, one for each \code{X} layer.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAAxisQueryResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAAxisQueryResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAAxisQueryResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
