% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMACollectionCreate}
\alias{SOMACollectionCreate}
\title{Create a SOMA Collection}
\usage{
SOMACollectionCreate(
  uri,
  ingest_mode = c("write", "resume"),
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL,
  soma_context = NULL
)
}
\arguments{
\item{uri}{URI for the TileDB object.}

\item{ingest_mode}{Ingestion mode when creating the TileDB object;
choose from:
\itemize{
\item \dQuote{\code{write}}: create a new TileDB object and error if
it already exists.
\item \dQuote{\code{resume}}: attempt to create a new TileDB object;
if it already exists, simply open it for writing.
}}

\item{platform_config}{Optional platform configuration.}

\item{tiledbsoma_ctx}{Optional (DEPRECATED) SOMATileDBContext.}

\item{tiledb_timestamp}{Optional Datetime (POSIXct) for TileDB timestamp.}

\item{soma_context}{Optional TileDB SOMA context.}
}
\value{
A new \link[tiledbsoma:SOMACollection]{SOMA collection} stored at
\code{uri} opened for writing.
}
\description{
Factory function to create a
\link[tiledbsoma:SOMACollection]{SOMA collection} for writing
(lifecycle: maturing).
}
\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
uri <- withr::local_tempfile(pattern = "soma-collection")

(col <- SOMACollectionCreate(uri))
col$add_new_sparse_ndarray("sparse", arrow::float64(), shape = c(100L, 100L))
col$close()

(col <- SOMACollectionOpen(uri))
col$names()

\dontshow{
col$close()
}
\dontshow{\}) # examplesIf}
}
