% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMADataFrame.R
\name{SOMADataFrame}
\alias{SOMADataFrame}
\title{SOMADataFrame}
\description{
\code{SOMADataFrame} is a multi-column table containing a "pseudo-column" called
\code{soma_rowid}, of type \code{uint64} and domain \verb{[0, #rows)}, which is the
row offset (row id), and is a contiguous integer beginning with zero.

\code{SOMADataFrame} must contain a column called \code{soma_joinid}, of type \code{uint64}.
The \code{soma_joinid} column contains a unique value for each row in the
\code{SOMADataFrame}, and intended to act as a joint key for other objects, such
as \code{\link{SOMASparseNdArray}}.
}
\section{Super classes}{
\code{tiledbsoma::TileDBObject} -> \code{tiledbsoma::TileDBArray} -> \code{SOMADataFrame}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMADataFrame-create}{\code{SOMADataFrame$create()}}
\item \href{#method-SOMADataFrame-write}{\code{SOMADataFrame$write()}}
\item \href{#method-SOMADataFrame-read}{\code{SOMADataFrame$read()}}
\item \href{#method-SOMADataFrame-clone}{\code{SOMADataFrame$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="class"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-class'><code>tiledbsoma::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="exists"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-exists'><code>tiledbsoma::TileDBObject$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="initialize"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-initialize'><code>tiledbsoma::TileDBObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="add_metadata"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-add_metadata'><code>tiledbsoma::TileDBArray$add_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="attributes"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-attributes'><code>tiledbsoma::TileDBArray$attributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="attrnames"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-attrnames'><code>tiledbsoma::TileDBArray$attrnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="dimensions"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-dimensions'><code>tiledbsoma::TileDBArray$dimensions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="dimnames"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-dimnames'><code>tiledbsoma::TileDBArray$dimnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="fragment_count"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-fragment_count'><code>tiledbsoma::TileDBArray$fragment_count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="get_metadata"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-get_metadata'><code>tiledbsoma::TileDBArray$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="print"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-print'><code>tiledbsoma::TileDBArray$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="reset_query"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-reset_query'><code>tiledbsoma::TileDBArray$reset_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="schema"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-schema'><code>tiledbsoma::TileDBArray$schema()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="set_query"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-set_query'><code>tiledbsoma::TileDBArray$set_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBArray" data-id="tiledb_array"><a href='../../tiledbsoma/html/TileDBArray.html#method-TileDBArray-tiledb_array'><code>tiledbsoma::TileDBArray$tiledb_array()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMADataFrame-create"></a>}}
\if{latex}{\out{\hypertarget{method-SOMADataFrame-create}{}}}
\subsection{Method \code{create()}}{
Create
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMADataFrame$create(schema)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{schema}}{an \code{\link[arrow:Schema]{arrow::schema}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMADataFrame-write"></a>}}
\if{latex}{\out{\hypertarget{method-SOMADataFrame-write}{}}}
\subsection{Method \code{write()}}{
Write
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMADataFrame$write(values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{An \code{\link[arrow:Table]{arrow::Table}} containing all columns, including
the \code{soma_rowid} index column. The schema for \code{values} must match the
schema for the \code{SOMADataFrame}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMADataFrame-read"></a>}}
\if{latex}{\out{\hypertarget{method-SOMADataFrame-read}{}}}
\subsection{Method \code{read()}}{
Read
Read a user-defined subset of data, addressed by the dataframe indexing
column, and optionally filtered.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMADataFrame$read(
  ids = NULL,
  column_names = NULL,
  value_filter = NULL,
  result_order = "UNORDERED"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{Integer indices specifying the rows to read.}

\item{\code{column_names}}{Character vector of column names to return.}

\item{\code{value_filter}}{A string containing a logical expression that is used
to filter the returned values. See \code{\link[tiledb:parse_query_condition]{tiledb::parse_query_condition}} for
more information.}

\item{\code{result_order}}{Order of read results. This can be one of either
\verb{"ROW_MAJOR, }"COL_MAJOR"\verb{, }"GLOBAL_ORDER"\verb{, or }"UNORDERED"`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{\link[arrow:Table]{arrow::Table}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMADataFrame-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMADataFrame-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMADataFrame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
