% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMADataFrameCreate}
\alias{SOMADataFrameCreate}
\title{Create SOMA DataFrame}
\usage{
SOMADataFrameCreate(uri, schema, index_column_names)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{schema}{schema Arrow schema argument passed on to DataFrame$create()}

\item{index_column_names}{Index column names passed on to DataFrame$create()}
}
\description{
Factory function to create a SOMADataFrame for writing, (lifecycle: experimental)
}
