% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMADenseNDArrayCreate}
\alias{SOMADenseNDArrayCreate}
\title{Create SOMA Dense Nd Array}
\usage{
SOMADenseNDArrayCreate(uri, type, shape)
}
\arguments{
\item{uri}{URI for the TileDB objec}

\item{type}{an \link[arrow:data-type]{Arrow type} defining the type of each element in the array.}

\item{shape}{a vector of integers defining the shape of the array.}
}
\description{
Factory function to create a SOMADenseNDArray for writing, (lifecycle: experimental)
}
