% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMADenseNDArrayOpen}
\alias{SOMADenseNDArrayOpen}
\title{Open SOMA Dense Nd Array}
\usage{
SOMADenseNDArrayOpen(uri, platform_config = NULL, tiledbsoma_ctx = NULL)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{Optional SOMATileDBContext}
}
\description{
Factory function to open a SOMADenseNDArray for reading, (lifecycle: experimental)
}
