% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAExperimentAxisQuery.R
\name{SOMAExperimentAxisQuery}
\alias{SOMAExperimentAxisQuery}
\title{\code{SOMAExperiment} Axis Query}
\description{
Perform an axis-based query against a \code{\link{SOMAExperiment}}.

\code{SOMAExperimentAxisQuery} allows easy selection and extraction of data from a
single \code{\link{SOMAMeasurement}} in a \code{\link{SOMAExperiment}}, by \code{obs}/\code{var} (axis)
coordinates and/or value filter. The primary use for this class is slicing
\code{\link{SOMAExperiment}} \code{X} layers by \code{obs} or \code{var} value and/or coordinates.
\link{lifecycle: experimental}
\subsection{X Layer Support}{

Slicing on \code{\link{SOMASparseNDArray}} \code{X} matrices is supported;
\code{\link{SOMADenseNDArray}} is not supported at this time.
}

\subsection{Result Size}{

\code{SOMAExperimentAxisQuery} query class assumes it can store the full result of
both axis dataframe queries in memory, and only provides incremental access
to the underlying X NDArray. Accessors such as \code{n_obs} and \code{n_vars} codify
this in the class.
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{experiment}}{The parent \code{\link{SOMAExperiment}} object.}

\item{\code{indexer}}{The \code{\link{SOMAIndexer}} object.}

\item{\code{obs_query}}{The \code{obs} \code{\link{SOMAAxisQuery}} object.}

\item{\code{var_query}}{The \code{var} \code{\link{SOMAAxisQuery}} object.}

\item{\code{n_obs}}{The number of \code{obs} axis query results.}

\item{\code{n_vars}}{The number of \code{var} axis query results.}

\item{\code{obs_df}}{The \code{obs} \code{\link{SOMADataFrame}} object.}

\item{\code{var_df}}{The \code{var} \code{\link{SOMADataFrame}} object for the specified
\code{measurement_name}.}

\item{\code{ms}}{The \code{\link{SOMAMeasurement}} object for the specified
\code{measurement_name}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMAExperimentAxisQuery-new}{\code{SOMAExperimentAxisQuery$new()}}
\item \href{#method-SOMAExperimentAxisQuery-obs}{\code{SOMAExperimentAxisQuery$obs()}}
\item \href{#method-SOMAExperimentAxisQuery-var}{\code{SOMAExperimentAxisQuery$var()}}
\item \href{#method-SOMAExperimentAxisQuery-obs_joinids}{\code{SOMAExperimentAxisQuery$obs_joinids()}}
\item \href{#method-SOMAExperimentAxisQuery-var_joinids}{\code{SOMAExperimentAxisQuery$var_joinids()}}
\item \href{#method-SOMAExperimentAxisQuery-X}{\code{SOMAExperimentAxisQuery$X()}}
\item \href{#method-SOMAExperimentAxisQuery-read}{\code{SOMAExperimentAxisQuery$read()}}
\item \href{#method-SOMAExperimentAxisQuery-clone}{\code{SOMAExperimentAxisQuery$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperimentAxisQuery-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperimentAxisQuery-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SOMAExperimentAxisQuery} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperimentAxisQuery$new(
  experiment,
  measurement_name,
  obs_query = NULL,
  var_query = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiment}}{A \code{\link{SOMAExperiment}} object.}

\item{\code{measurement_name}}{The name of the measurement to query.}

\item{\code{obs_query, var_query}}{An \code{\link{SOMAAxisQuery}} object for the obs/var
axis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperimentAxisQuery-obs"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperimentAxisQuery-obs}{}}}
\subsection{Method \code{obs()}}{
Retrieve obs \code{\link[arrow:Table]{arrow::Table}}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperimentAxisQuery$obs(column_names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column_names}}{A character vector of column names to retrieve}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperimentAxisQuery-var"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperimentAxisQuery-var}{}}}
\subsection{Method \code{var()}}{
Retrieve var \code{\link[arrow:Table]{arrow::Table}}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperimentAxisQuery$var(column_names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column_names}}{A character vector of column names to retrieve}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperimentAxisQuery-obs_joinids"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperimentAxisQuery-obs_joinids}{}}}
\subsection{Method \code{obs_joinids()}}{
Retrieve \code{soma_joinids} as an \code{\link[arrow:array]{arrow::Array}} for \code{obs}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperimentAxisQuery$obs_joinids()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperimentAxisQuery-var_joinids"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperimentAxisQuery-var_joinids}{}}}
\subsection{Method \code{var_joinids()}}{
Retrieve \code{soma_joinids} as an \code{\link[arrow:array]{arrow::Array}} for \code{var}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperimentAxisQuery$var_joinids()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperimentAxisQuery-X"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperimentAxisQuery-X}{}}}
\subsection{Method \code{X()}}{
Retrieves an \code{X} layer as an \code{\link[arrow:Table]{arrow::Table}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperimentAxisQuery$X(layer_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layer_name}}{The name of the layer to retrieve.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperimentAxisQuery-read"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperimentAxisQuery-read}{}}}
\subsection{Method \code{read()}}{
Reads the entire query result as a list of
\code{\link[arrow:Table]{arrow::Table}}s. This is a low-level routine intended to be used by
loaders for other in-core formats, such as \code{Seurat}, which can be created
from the resulting Tables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperimentAxisQuery$read(
  X_layers = NULL,
  obs_column_names = NULL,
  var_column_names = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_layers}}{The name of the \code{X} layer(s) to read and return.}

\item{\code{obs_column_names, var_column_names}}{Specify which column names in
\code{var} and \code{obs} dataframes to read and return.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperimentAxisQuery-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperimentAxisQuery-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperimentAxisQuery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
