% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMAExperimentCreate}
\alias{SOMAExperimentCreate}
\title{Create SOMA Experiment}
\usage{
SOMAExperimentCreate(uri, platform_config = NULL, tiledbsoma_ctx = NULL)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{Optional SOMATileDBContext}
}
\description{
Factory function to create a SOMADataFrame for writing, (lifecycle: experimental)
}
