% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMAMeasurementCreate}
\alias{SOMAMeasurementCreate}
\title{Create a SOMA Measurement}
\usage{
SOMAMeasurementCreate(
  uri,
  ingest_mode = c("write", "resume"),
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL,
  context = NULL
)
}
\arguments{
\item{uri}{URI for the TileDB object.}

\item{ingest_mode}{Ingestion mode when creating the TileDB object;
choose from:
\itemize{
\item \dQuote{\code{write}}: create a new TileDB object and error if
it already exists.
\item \dQuote{\code{resume}}: attempt to create a new TileDB object;
if it already exists, simply open it for writing.
}}

\item{platform_config}{Optional platform configuration.}

\item{tiledbsoma_ctx}{Optional (DEPRECATED) SOMATileDBContext.}

\item{tiledb_timestamp}{Optional Datetime (POSIXct) for TileDB timestamp.}

\item{context}{Optional \code{SOMAContext} object used for TileDB operations.
If a context is not provided, then the default context will be used. Call
\code{set_default_context} once before other SOMA operations to configure
the default context.}
}
\value{
A new \link[tiledbsoma:SOMAMeasurement]{SOMA measurement} stored at
\code{uri} opened for writing.
}
\description{
Factory function to create a
\link[tiledbsoma:SOMAMeasurement]{SOMA measurement} for writing
(lifecycle: maturing).
}
\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
uri <- withr::local_tempfile(pattern = "soma-measurement")
var <- data.frame(
  soma_joinid = bit64::seq.integer64(0L, 99L),
  var_id = paste0("feature_", seq_len(100L))
)
sch <- arrow::infer_schema(var)

(ms <- SOMAMeasurementCreate(uri))
sdf <- ms$add_new_dataframe(
  "var",
  sch,
  "soma_joinid",
  list(soma_joinid = c(0, 100))
)
sdf$write(arrow::as_arrow_table(var, schema = sch))
sdf$close()
ms$close()

(ms <- SOMAMeasurementOpen(uri))
ms$var

\dontshow{
ms$close()
}
\dontshow{\}) # examplesIf}
}
