% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAObject.R
\name{SOMAObject}
\alias{SOMAObject}
\title{The SOMA Object Base Class}
\description{
Base class to implement shared functionality across the
\code{\link{SOMAArrayBase}} and \code{\link{SOMACollectionBase}} classes.
(lifecycle: maturing)
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{platform_config}}{Platform configuration}

\item{\code{tiledbsoma_ctx}}{SOMATileDBContext}

\item{\code{tiledb_timestamp}}{Time that object was opened at}

\item{\code{uri}}{The URI of the TileDB object.}

\item{\code{.tiledb_timestamp_range}}{Time range for libtiledbsoma}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMAObject-new}{\code{SOMAObject$new()}}
\item \href{#method-SOMAObject-reopen}{\code{SOMAObject$reopen()}}
\item \href{#method-SOMAObject-clone}{\code{SOMAObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-new}{}}}
\subsection{Method \code{new()}}{
Create a new SOMA object. (lifecycle: maturing)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$new(
  uri,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL,
  internal_use_only = NULL,
  soma_context = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{...}

\item{\code{platform_config}}{...}

\item{\code{tiledbsoma_ctx}}{...}

\item{\code{tiledb_timestamp}}{...}

\item{\code{internal_use_only}}{...}

\item{\code{soma_context}}{...}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-reopen"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-reopen}{}}}
\subsection{Method \code{reopen()}}{
Close and reopen the TileDB object in a new mode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$reopen(mode, tiledb_timestamp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{New mode to open the object in; choose from:
\itemize{
\item \dQuote{\code{READ}}
\item \dQuote{\code{WRITE}}
}}

\item{\code{tiledb_timestamp}}{Optional Datetime (POSIXct) with TileDB timestamp}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns \code{self} opened in \code{mode}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
