% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAOpen.R
\name{SOMAOpen}
\alias{SOMAOpen}
\title{Open a SOMA Object}
\usage{
SOMAOpen(uri, platform_config = NULL, tiledbsoma_ctx = NULL)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{Optional SOMATileDBContext}
}
\description{
Utility function to open the corresponding SOMA Object given a URI, (lifecycle: experimental)
}
