% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMASparseNDArrayRead.R
\name{SOMASparseNDArrayBlockwiseRead}
\alias{SOMASparseNDArrayBlockwiseRead}
\title{Blockwise Sparse ND-Array Reader}
\description{
Blockwise reader for \code{\link{SOMASparseNDArray}}
}
\keyword{internal}
\section{Super class}{
\code{\link[tiledbsoma:SOMASparseNDArrayReadBase]{tiledbsoma::SOMASparseNDArrayReadBase}} -> \code{SOMASparseNDArrayBlockwiseRead}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{axis}}{...}

\item{\code{size}}{...}

\item{\code{reindex_disable_on_axis}}{...}

\item{\code{eager}}{...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMASparseNDArrayBlockwiseRead-new}{\code{SOMASparseNDArrayBlockwiseRead$new()}}
\item \href{#method-SOMASparseNDArrayBlockwiseRead-tables}{\code{SOMASparseNDArrayBlockwiseRead$tables()}}
\item \href{#method-SOMASparseNDArrayBlockwiseRead-sparse_matrix}{\code{SOMASparseNDArrayBlockwiseRead$sparse_matrix()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayBlockwiseRead-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayBlockwiseRead-new}{}}}
\subsection{Method \code{new()}}{
Create
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayBlockwiseRead$new(
  sr,
  array,
  coords,
  axis,
  ...,
  size,
  reindex_disable_on_axis = NULL,
  eager = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sr}}{SOMA read pointer}

\item{\code{array}}{\code{\link{SOMASparseNDArray}}}

\item{\code{coords}}{...}

\item{\code{axis}}{...}

\item{\code{...}}{Ignored}

\item{\code{size}}{...}

\item{\code{reindex_disable_on_axis}}{...}

\item{\code{eager}}{...}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayBlockwiseRead-tables"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayBlockwiseRead-tables}{}}}
\subsection{Method \code{tables()}}{
...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayBlockwiseRead$tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayBlockwiseRead-sparse_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayBlockwiseRead-sparse_matrix}{}}}
\subsection{Method \code{sparse_matrix()}}{
...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayBlockwiseRead$sparse_matrix(compress = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{compress}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
...
}
}
}
