% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMASparseNDArrayRead.R
\name{SOMASparseNDArrayRead}
\alias{SOMASparseNDArrayRead}
\title{SOMASparseNDArrayRead}
\description{
Intermediate type to choose result format when reading a sparse array
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMASparseNDArrayRead-new}{\code{SOMASparseNDArrayRead$new()}}
\item \href{#method-SOMASparseNDArrayRead-sparse_matrix}{\code{SOMASparseNDArrayRead$sparse_matrix()}}
\item \href{#method-SOMASparseNDArrayRead-tables}{\code{SOMASparseNDArrayRead$tables()}}
\item \href{#method-SOMASparseNDArrayRead-clone}{\code{SOMASparseNDArrayRead$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayRead-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayRead-new}{}}}
\subsection{Method \code{new()}}{
Create (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayRead$new(sr, shape)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sr}}{soma read pointer}

\item{\code{shape}}{Shape of the full matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayRead-sparse_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayRead-sparse_matrix}{}}}
\subsection{Method \code{sparse_matrix()}}{
Read as a sparse matrix (lifecycle: experimental). Returns
an iterator of Matrix::\link[Matrix]{dgTMatrix-class} or \link{matrixZeroBasedView} of it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayRead$sparse_matrix(zero_based = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zero_based}}{Logical, if \code{TRUE} returns iterator of \link{matrixZeroBasedView}
if \code{FALSE} returns iterator of Matrix::\link[Matrix]{dgTMatrix-class}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{SparseReadIter}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayRead-tables"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayRead-tables}{}}}
\subsection{Method \code{tables()}}{
Read as a arrow::\link[arrow]{Table} (lifecycle: experimental).
Returns an iterator of arrow::\link[arrow]{Table}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayRead$tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{TableReadIter}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayRead-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayRead-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayRead$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
