% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMASparseNDArrayRead.R
\name{SOMASparseNDArrayReadBase}
\alias{SOMASparseNDArrayReadBase}
\title{SOMA Sparse ND-Array Reader Base}
\description{
Base class for SOMA sparse ND-array reads
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sr}}{The SOMA read pointer}

\item{\code{array}}{The underlying \code{\link{SOMASparseNDArray}}}

\item{\code{coords}}{The coordinates for the read}

\item{\code{shape}}{The shape of the underlying array}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMASparseNDArrayReadBase-new}{\code{SOMASparseNDArrayReadBase$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayReadBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayReadBase-new}{}}}
\subsection{Method \code{new()}}{
Create
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayReadBase$new(sr, array, coords)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sr}}{SOMA read pointer}

\item{\code{array}}{\code{\link{SOMASparseNDArray}}}

\item{\code{coords}}{...}
}
\if{html}{\out{</div>}}
}
}
}
