% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{[.matrixZeroBasedView}
\alias{[.matrixZeroBasedView}
\title{Zero-based matrix element access}
\usage{
\method{[}{matrixZeroBasedView}(x, i, j, drop = TRUE)
}
\arguments{
\item{x}{The zero-based matrix view.}

\item{i}{Row index (zero-based).}

\item{j}{Column index (zero-based).}

\item{drop}{Coerce result to lowest possible dimension.}
}
\value{
The specified matrix elements. Vector- or matrix-valued results are returned
as conventional one-based R objects.
}
\description{
Zero-based matrix element access
}
