% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{[<-.matrixZeroBasedView}
\alias{[<-.matrixZeroBasedView}
\title{Zero-based matrix element assigment}
\usage{
\method{[}{matrixZeroBasedView}(x, i, j) <- value
}
\arguments{
\item{x}{The zero-based matrix view.}

\item{i}{Row index (zero-based).}

\item{j}{Column index (zero-based).}

\item{val}{The to-be assigned value.}
}
\value{
Nothing as the method errors.
}
\description{
This method errors as a read-only view is implemented.
}
