% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableReadIter.R
\name{TableReadIter}
\alias{TableReadIter}
\title{SOMA Read Iterator over Arrow Table}
\description{
\code{TableReadIter} is a class that allows for iteration over
a reads on \link{SOMASparseNDArray} and \link{SOMADataFrame}.
Iteration chunks are retrieved as arrow::\link[arrow]{Table}
}
\section{Super class}{
\code{\link[tiledbsoma:ReadIter]{tiledbsoma::ReadIter}} -> \code{TableReadIter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableReadIter-concat}{\code{TableReadIter$concat()}}
\item \href{#method-TableReadIter-clone}{\code{TableReadIter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="initialize"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-initialize'><code>tiledbsoma::ReadIter$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="read_complete"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-read_complete'><code>tiledbsoma::ReadIter$read_complete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="read_next"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-read_next'><code>tiledbsoma::ReadIter$read_next()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableReadIter-concat"></a>}}
\if{latex}{\out{\hypertarget{method-TableReadIter-concat}{}}}
\subsection{Method \code{concat()}}{
Concatenate remainder of iterator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableReadIter$concat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
arrow::\link[arrow]{Table}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableReadIter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableReadIter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableReadIter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
