% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tiledb_embedded_version}
\alias{tiledb_embedded_version}
\title{TileDB Embedded Version interface}
\usage{
tiledb_embedded_version()
}
\description{
This gets the version of the TileDB Embedded library that is currently in use.
}
