% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBGroup.R
\name{TileDBGroup}
\alias{TileDBGroup}
\title{TileDB Group Base Class}
\value{
A \code{TileDBArray} or \code{TileDBGroup}.
}
\description{
Base class for interacting with TileDB groups (lifecycle: experimental)
}
\keyword{internal}
\section{Super class}{
\code{\link[tiledbsoma:TileDBObject]{tiledbsoma::TileDBObject}} -> \code{TileDBGroup}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TileDBGroup-print}{\code{TileDBGroup$print()}}
\item \href{#method-TileDBGroup-create}{\code{TileDBGroup$create()}}
\item \href{#method-TileDBGroup-open}{\code{TileDBGroup$open()}}
\item \href{#method-TileDBGroup-close}{\code{TileDBGroup$close()}}
\item \href{#method-TileDBGroup-set}{\code{TileDBGroup$set()}}
\item \href{#method-TileDBGroup-get}{\code{TileDBGroup$get()}}
\item \href{#method-TileDBGroup-remove}{\code{TileDBGroup$remove()}}
\item \href{#method-TileDBGroup-length}{\code{TileDBGroup$length()}}
\item \href{#method-TileDBGroup-names}{\code{TileDBGroup$names()}}
\item \href{#method-TileDBGroup-to_list}{\code{TileDBGroup$to_list()}}
\item \href{#method-TileDBGroup-to_data_frame}{\code{TileDBGroup$to_data_frame()}}
\item \href{#method-TileDBGroup-get_metadata}{\code{TileDBGroup$get_metadata()}}
\item \href{#method-TileDBGroup-set_metadata}{\code{TileDBGroup$set_metadata()}}
\item \href{#method-TileDBGroup-clone}{\code{TileDBGroup$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="class"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-class'><code>tiledbsoma::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="exists"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-exists'><code>tiledbsoma::TileDBObject$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="initialize"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-initialize'><code>tiledbsoma::TileDBObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="is_open"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-is_open'><code>tiledbsoma::TileDBObject$is_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="mode"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-mode'><code>tiledbsoma::TileDBObject$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="reopen"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-reopen'><code>tiledbsoma::TileDBObject$reopen()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-print"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-print}{}}}
\subsection{Method \code{print()}}{
Print summary of the group. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-create"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-create}{}}}
\subsection{Method \code{create()}}{
Creates the data structure on disk/S3/cloud. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$create(internal_use_only = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{internal_use_only}}{Character value to signal this is a 'permitted' call,
as \code{create()} is considered internal and should not be called directly.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-open"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-open}{}}}
\subsection{Method \code{open()}}{
Open the SOMA object for read or write.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$open(mode = c("READ", "WRITE"), internal_use_only = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{Mode to open in; defaults to \code{READ}.}

\item{\code{internal_use_only}}{Character value to signal this is a 'permitted' call,
as \code{open()} is considered internal and should not be called directly.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-close"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-close}{}}}
\subsection{Method \code{close()}}{
Close the SOMA object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-set"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-set}{}}}
\subsection{Method \code{set()}}{
Add new member to the group. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$set(object, name = NULL, relative = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{TileDBArray} or \code{TileDBGroup} object to add.}

\item{\code{name}}{Name to use for the member. By default the base name of
the object's URI is used.}

\item{\code{relative}}{An optional logical value indicating whether the new
object's URI is relative to the group's URI. If \code{NULL} (the
default), the object's URI is assumed to be relative unless it is a
\verb{tiledb://} URI.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-get"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-get}{}}}
\subsection{Method \code{get()}}{
Retrieve a group member by name. If the member isn't already
open, it is opened in the same mode as the parent. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the member.}

\item{\code{mode}}{Mode to open in}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-remove"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-remove}{}}}
\subsection{Method \code{remove()}}{
Remove member. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$remove(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the member to remove.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-length"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-length}{}}}
\subsection{Method \code{length()}}{
Length in the number of members. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$length()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Scalar \code{integer}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-names"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-names}{}}}
\subsection{Method \code{names()}}{
Retrieve the names of members. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} vector of member names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-to_list}{}}}
\subsection{Method \code{to_list()}}{
Retrieve a \code{list} of members. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-to_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-to_data_frame}{}}}
\subsection{Method \code{to_data_frame()}}{
Retrieve a \code{data.frame} of members. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$to_data_frame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Retrieve metadata. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_metadata(key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The name of the metadata attribute to retrieve.
is not NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of metadata values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-set_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-set_metadata}{}}}
\subsection{Method \code{set_metadata()}}{
Add list of metadata. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$set_metadata(metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Named list of metadata to add.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
