% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledbsc-package.R
\docType{package}
\name{tiledbsc-package}
\alias{tiledbsc}
\alias{tiledbsc-package}
\title{tiledbsc: TileDB Single Cell SOMA API}
\description{
Interface for working with 'TileDB'-based Stack of Matrices, Annotated ('SOMA'): an open data model for representing annotated matrices, like those commonly used for single cell data analysis. Users can import from and export to in-memory formats used by popular toolchains like 'Seurat', 'Bioconductor', and even 'AnnData' using the companion Python package.
}
\details{
The API is designed around R6-based classes corresponding to the
components described in the \href{https://github.com/single-cell-data/soma}{data model specification}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TileDB-Inc/tiledbsc}
  \item \url{https://github.com/single-cell-data/SOMA}
  \item Report bugs at \url{https://github.com/TileDB-Inc/tiledbsc/issues}
}

}
\author{
\strong{Maintainer}: Aaron Wolen \email{aaron@tiledb.com} (\href{https://orcid.org/0000-0003-2542-2202}{ORCID})

Authors:
\itemize{
  \item Dirk Eddelbuettel \email{dirk@tiledb.com} (\href{https://orcid.org/0000-0001-6419-907X}{ORCID})
  \item John Kerl \email{john.kerl@tiledb.com}
}

Other contributors:
\itemize{
  \item TileDB, Inc. [copyright holder, funder]
}

}
\keyword{internal}
