% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.R
\name{tiledbsoma_empty_query_condition}
\alias{tiledbsoma_empty_query_condition}
\title{Creates a 'tiledbsoma_query_condition' object}
\usage{
tiledbsoma_empty_query_condition(somactx)
}
\arguments{
\item{somactx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A 'tiledbsoma_query_condition' object
}
\description{
Creates a 'tiledbsoma_query_condition' object
}
