% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utils-tiledb.R
\name{tiledbsoma_stats}
\alias{tiledbsoma_stats}
\alias{tiledbsoma_stats_enable}
\alias{tiledbsoma_stats_disable}
\alias{tiledbsoma_stats_reset}
\alias{tiledbsoma_stats_dump}
\alias{tiledbsoma_stats_show}
\title{TileDB SOMA Statistics}
\usage{
tiledbsoma_stats_enable()

tiledbsoma_stats_disable()

tiledbsoma_stats_reset()

tiledbsoma_stats_dump()

tiledbsoma_stats_show()
}
\value{
\code{tiledbsoma_stats_show()}: a single-length character vector
with the TileDB statistics encoded in JSON format

All other functions invisibly return \code{NULL}
}
\description{
These functions expose the TileDB Core functionality for performance
measurements and statistics
}
\details{
\itemize{
\item \code{tiledbsoma_stats_enable()}/\code{tiledbsoma_stats_disable()}:
Enable and disable TielDB's internal statistics
\item \code{tiledbsoma_stats_reset()}: Reset all statistics to \code{0}
\item \code{tiledbsoma_stats_dump()}: Dump all statistics as a JSON string
\item \code{tiledbsoma_stats_show()}: Pretty-print the JSON statistics
}
}
