% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seurat.R
\name{write_soma.Seurat}
\alias{write_soma.Seurat}
\title{Write a \code{\link[SeuratObject]{Seurat}} object to a SOMA, returned opened for write}
\usage{
\method{write_soma}{Seurat}(x, uri, ..., platform_config = NULL, tiledbsoma_ctx = NULL)
}
\arguments{
\item{x}{A \code{\link[SeuratObject]{Seurat}} object}

\item{uri}{URI for resulting SOMA object}

\item{...}{Arguments passed to other methods}

\item{platform_config}{Optional \link[tiledbsoma:PlatformConfig]{platform
configuration}}

\item{tiledbsoma_ctx}{Optional \code{\link{SOMATileDBContext}}}
}
\value{
The URI to the resulting \code{\link{SOMAExperiment}} generated from
the data contained in \code{x}
}
\description{
Write a \code{\link[SeuratObject]{Seurat}} object to a SOMA, returned opened for write
}
\section{Writing Cell-Level Meta Data}{

Cell-level meta data is written out as a
\link[tiledbsoma:SOMADataFrame]{data frame} called \dQuote{\code{obs}} at
the \code{\link[tiledbsoma:SOMAExperiment]{experiment}} level
}

\section{Writing \code{\link[SeuratObject]{Assay}s}}{

\pkg{Seurat} \code{\link[SeuratObject]{Assay}} objects are written out as
individual \link[tiledbsoma:SOMAMeasurement]{measurements}:
\itemize{
\item the \dQuote{\code{data}} matrix is written out a
\link[tiledbsoma:SOMASparseNDArray]{sparse matrix} called
\dQuote{\code{data}} within the \dQuote{\code{X}} group
\item the \dQuote{\code{counts}} matrix, if not
\link[SeuratObject:IsMatrixEmpty]{empty}, is written out a
\link[tiledbsoma:SOMASparseNDArray]{sparse matrix} called
\dQuote{\code{counts}} within the \dQuote{\code{X}} group
\item the \dQuote{\code{scale.data}} matrix, if not
\link[SeuratObject:IsMatrixEmpty]{empty}, is written out a
\link[tiledbsoma:SOMASparseNDArray]{sparse matrix} called
\dQuote{\code{scale_data}} within the \dQuote{\code{X}} group
\item feature-level meta data is written out as a
\link[tiledbsoma:SOMADataFrame]{data frame} called \dQuote{\code{var}}
}
Expression matrices are transposed (cells as rows) prior to writing. All
other slots, including results from extended assays (eg. \code{SCTAssay},
\code{ChromatinAssay}) are lost
}

\section{Writing \code{\link[SeuratObject]{DimReduc}s}}{

\pkg{Seurat} \code{\link[SeuratObject]{DimReduc}} objects are written out
to the \dQuote{\code{obsm}} and \dQuote{\code{varm}} groups of a
\link[tiledbsoma:SOMAMeasurement]{measurement}:
\itemize{
\item cell embeddings are written out as a
\link[tiledbsoma:SOMASparseNDArray]{sparse matrix} in the
\dQuote{\code{obsm}} group
\item feature loadings, if not \link[SeuratObject:IsMatrixEmpty]{empty},
are written out as a \link[tiledbsoma:SOMASparseNDArray]{sparse matrix} in
the \dQuote{\code{varm}} groups; loadings are padded with \code{NAs}
to include all features
}
Dimensional reduction names are translated to AnnData-style names (eg.
\dQuote{\code{pca}} becomes \code{X_pca} for embeddings and
\dQuote{\code{PCs}} for loadings). All other slots, including projected
feature loadings and jackstraw information, are lost
}

\section{Writing \code{\link[SeuratObject]{Graph}s}}{

\pkg{Seurat} \code{\link[SeuratObject]{Graph}} objects are
written out as \link[tiledbsoma:SOMASparseNDArray]{sparse matrices}
to the \dQuote{\code{obsp}} group of a
\link[tiledbsoma:SOMAMeasurement]{measurement}
}

