% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bioc.R
\name{write_soma.SingleCellExperiment}
\alias{write_soma.SingleCellExperiment}
\title{Write a \code{\link[SingleCellExperiment:SingleCellExperiment-class]{SingleCellExperiment}}
object to a SOMA}
\usage{
\method{write_soma}{SingleCellExperiment}(
  x,
  uri,
  ms_name = NULL,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL
)
}
\arguments{
\item{x}{An object}

\item{uri}{URI for resulting SOMA object}

\item{ms_name}{Name for resulting measurement; defaults to
\code{\link[SingleCellExperiment]{mainExpName}(x)}}

\item{...}{Arguments passed to other methods}

\item{platform_config}{Optional \link[tiledbsoma:PlatformConfig]{platform
configuration}}

\item{tiledbsoma_ctx}{Optional \code{\link{SOMATileDBContext}}}
}
\value{
The URI to the resulting \code{\link{SOMAExperiment}} generated from
the data contained in \code{x}, returned opened for write
}
\description{
Write a \code{\link[SingleCellExperiment:SingleCellExperiment-class]{SingleCellExperiment}}
object to a SOMA
}
\section{Writing \code{colData}}{

\code{colData} is written out as a
\link[tiledbsoma:SOMADataFrame]{data frame} called \dQuote{\code{obs}} at
the \code{\link[tiledbsoma:SOMAExperiment]{experiment}} level
}

\section{Writing Assay Matrices}{

Each \link[SummarizedExperiment:assay]{assay matrix} is written out as a
\link[tiledbsoma:SOMASparseNDArray]{sparse matrix} within the \code{X} group of
\code{\link[tiledbsoma:SOMAMeasurement]{measurement}} names \code{ms_name}. Names
for assay matrices within \code{X} are taken from the
\link[SummarizedExperiment:assayNames]{assay names}. Assay matrices are
transposed (samples as rows) prior to writing
}

\section{Writing \code{rowData}}{

\code{rowData} is written out as a
\link[tiledbsoma:SOMADataFrame]{data frame} called \dQuote{\code{var}} at
the \code{\link[tiledbsoma:SOMAMeasurement]{measurement}} level
}

