% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_soma.R
\name{write_soma_objects}
\alias{write_soma_objects}
\alias{write_soma.data.frame}
\alias{write_soma.matrix}
\alias{write_soma.Matrix}
\alias{write_soma.TsparseMatrix}
\title{Write R Objects to SOMA}
\usage{
\method{write_soma}{data.frame}(
  x,
  uri,
  soma,
  index = NULL,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  absolute = FALSE
)

\method{write_soma}{matrix}(
  x,
  uri,
  soma,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  absolute = FALSE
)

\method{write_soma}{Matrix}(
  x,
  uri,
  soma,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  absolute = FALSE
)

\method{write_soma}{TsparseMatrix}(
  x,
  uri,
  soma,
  type = NULL,
  transpose = FALSE,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  absolute = FALSE
)
}
\arguments{
\item{x}{An object}

\item{uri}{URI for resulting SOMA}

\item{soma}{The parent \link[tiledbsoma:SOMACollection]{collection} (eg. a
\code{\link{SOMACollection}}, \code{\link{SOMAExperiment}}, or
\code{\link{SOMAMeasurement}})}

\item{index}{The name of the column in \code{x} with the index (row names);
by default, will automatically add the row names of \code{x} to \code{x}
before writing}

\item{...}{Arguments passed to other methods}

\item{platform_config}{Optional \link[tiledbsoma:PlatformConfig]{platform
configuration}}

\item{tiledbsoma_ctx}{Optional \code{\link{SOMATileDBContext}}}

\item{absolute}{\strong{[Internal use only]} Is \code{uri} absolute}

\item{sparse}{Create a \link[tiledbsoma:SOMASparseNDArray]{sparse} or
\link[tiledbsoma:SOMADenseNDArray]{dense} array from \code{x}}

\item{type}{\link[arrow:data-type]{Arrow type} for encoding \code{x}
(eg. \code{\link[arrow:data-type]{arrow::int32}()}); by default, attempts to
determine arrow type with \code{\link[arrow:infer_type]{arrow::infer_type}()}}

\item{transpose}{Transpose \code{x} before writing}
}
\value{
The resulting SOMA \link[tiledbsoma:SOMASparseNDArray]{array} or
\link[tiledbsoma:SOMADataFrame]{data frame}
}
\description{
Various helpers to write R objects to SOMA
}
\section{Writing Data Frames}{

\link[base:data.frame]{Data frames} are written out as
\link[tiledbsoma:SOMADataFrame]{SOMA-structured TileDB data frames}. The
following transformations are applied to \code{x}:
\itemize{
\item row names are added to a column in \code{x} entitled
\dQuote{\code{_index}}, \dQuote{\code{rownames}}, or a random name if
either option is already present in \code{x}
\item a column \dQuote{\code{soma_joinid}} will be automatically added
going from \code{[0, nrow(x) - 1]} encoded as
\link[bit64:integer64]{64-bit integers}; any columns called
\dQuote{\code{soma_joinid}} will be overwritten if present in \code{x}
\item all factor columns will be coerced to characters
\item all columns not containing \link[base:is.atomic]{atomic} types
(excluding \link[base:factor]{factors}, \code{\link[base]{complex}es},
and \code{\link[base]{raw}s}) will be removed
}
The array type for each column will be determined by
\code{\link[arrow:infer_type]{arrow::infer_type}()}
}

\section{Writing Dense Matrices}{

Dense matrices are written as two-dimensional
\link[tiledbsoma:SOMADenseNDArray]{dense arrays}. The overall shape of the
array is determined by \code{dim(x)} and the type of the array is determined
by \code{type} or \code{\link[arrow:infer_type]{arrow::infer_type}(x)}
}

\section{Writing Sparse Matrices}{

Sparse matrices are written out as two-dimensional
\link[tiledbsoma:SOMASparseNDArray]{TileDB sparse arrays} in
\href{https://en.wikipedia.org/wiki/Sparse_matrix#Coordinate_list_(COO)}{COO format}:
\itemize{
\item the row indices (\dQuote{\code{i}}) are written out as \dQuote{\code{soma_dim_0}}
\item the column indices (\dQuote{\code{j}}) are written out as \dQuote{\code{soma_dim_1}}
\item the non-zero values (\dQuote{\code{x}}) are written out as \dQuote{\code{soma_data}}
}
The array type is determined by \code{type}, or
\code{\link[arrow:infer_type]{arrow::infer_type}(slot(x, "x"))}
}

\keyword{internal}
