% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seurat.R
\name{write_soma_seurat_sub_objects}
\alias{write_soma_seurat_sub_objects}
\alias{write_soma.Assay}
\alias{write_soma.DimReduc}
\alias{write_soma.Graph}
\title{Write \pkg{Seurat} Sub-Objects to a SOMA}
\usage{
\method{write_soma}{Assay}(
  x,
  uri = NULL,
  soma,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  absolute = FALSE
)

\method{write_soma}{DimReduc}(
  x,
  uri = NULL,
  soma,
  fidx = NULL,
  nfeatures = NULL,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  absolute = FALSE
)

\method{write_soma}{Graph}(
  x,
  uri = NULL,
  soma,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  absolute = FALSE
)
}
\arguments{
\item{x}{A \pkg{Seurat} sub-object (eg. an
\code{\link[SeuratObject]{Assay}}, a \code{\link[SeuratObject]{DimReduc}},
or a \code{\link[SeuratObject]{Graph}})}

\item{uri}{URI for resulting SOMA}

\item{soma}{The parent \link[tiledbsoma:SOMACollection]{collection}; for
the \code{DimReduc} and \code{Graph} methods, this \strong{must} be a
\link[tiledbsoma:SOMAMeasurement]{measurement} for the assay \code{x}
was generated from}

\item{...}{Arguments passed to other methods}

\item{platform_config}{Optional \link[tiledbsoma:PlatformConfig]{platform
configuration}}

\item{tiledbsoma_ctx}{Optional \code{\link{SOMATileDBContext}}}

\item{absolute}{\strong{[Internal use only]} Is \code{uri} absolute}

\item{fidx}{An integer vector describing the location of features in
\code{SeuratObject::Loadings(x)} with relation to \code{soma}
(eg. \code{match(rownames(Loadings(x)), rownames(assay))})}

\item{nfeatures}{The number of features present in \code{soma}
(eg. \code{nrow(assay)})}
}
\value{
\code{Assay} method: a \code{\link{SOMAMeasurement}} with the
data from \code{x}

\code{DimReduc} and \code{Graph} methods: invisibly returns
\code{soma} with the values of \code{x} added to it
}
\description{
Various helpers to write \pkg{Seurat} sub-objects to SOMA
}
\section{Writing \code{\link[SeuratObject]{Assay}s}}{

\pkg{Seurat} \code{\link[SeuratObject]{Assay}} objects are written out as
individual \link[tiledbsoma:SOMAMeasurement]{measurements}:
\itemize{
\item the \dQuote{\code{data}} matrix is written out a
\link[tiledbsoma:SOMASparseNDArray]{sparse matrix} called
\dQuote{\code{data}} within the \dQuote{\code{X}} group
\item the \dQuote{\code{counts}} matrix, if not
\link[SeuratObject:IsMatrixEmpty]{empty}, is written out a
\link[tiledbsoma:SOMASparseNDArray]{sparse matrix} called
\dQuote{\code{counts}} within the \dQuote{\code{X}} group
\item the \dQuote{\code{scale.data}} matrix, if not
\link[SeuratObject:IsMatrixEmpty]{empty}, is written out a
\link[tiledbsoma:SOMASparseNDArray]{sparse matrix} called
\dQuote{\code{scale_data}} within the \dQuote{\code{X}} group
\item feature-level meta data is written out as a
\link[tiledbsoma:SOMADataFrame]{data frame} called \dQuote{\code{var}}
}
Expression matrices are transposed (cells as rows) prior to writing. All
other slots, including results from extended assays (eg. \code{SCTAssay},
\code{ChromatinAssay}) are lost
}

\section{Writing \code{\link[SeuratObject]{DimReduc}s}}{

\pkg{Seurat} \code{\link[SeuratObject]{DimReduc}} objects are written out
to the \dQuote{\code{obsm}} and \dQuote{\code{varm}} groups of a
\link[tiledbsoma:SOMAMeasurement]{measurement}:
\itemize{
\item cell embeddings are written out as a
\link[tiledbsoma:SOMADenseNDArray]{dense matrix} in the
\dQuote{\code{obsm}} group
\item feature loadings, if not \link[SeuratObject:IsMatrixEmpty]{empty},
are written out as a \link[tiledbsoma:SOMADenseNDArray]{dense matrix} in
the \dQuote{\code{varm}} groups; loadings are padded include all features
with \code{NAs}
}
Dimensional reduction names are translated to AnnData-style names (eg.
\dQuote{\code{pca}} becomes \code{X_pca} for embeddings and
\dQuote{\code{PCs}} for loadings). All other slots, including projected
feature loadings and jackstraw information, are lost
}

\section{Writing \code{\link[SeuratObject]{Graph}s}}{

\pkg{Seurat} \code{\link[SeuratObject]{Graph}} objects are
written out as \link[tiledbsoma:SOMASparseNDArray]{sparse matrices}
to the \dQuote{\code{obsp}} group of a
\link[tiledbsoma:SOMAMeasurement]{measurement}
}

\keyword{internal}
