% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-readerTransformers.R
\name{arrow_table_to_sparse}
\alias{arrow_table_to_sparse}
\title{Transformer function: Arrow table to Matrix::sparseMatrix}
\usage{
arrow_table_to_sparse(
  tbl,
  repr = c("C", "T", "R"),
  shape = NULL,
  zero_based = FALSE
)
}
\arguments{
\item{tbl}{\link[arrow]{Table} with columns "soma_dim_0", "soma_dim_1", and "soma_datah"}

\item{repr}{Optional one-character code for sparse matrix representation type}

\item{shape}{Numerical vector with two elements, one for each dimension. If
\code{NULL}, then the following is used \code{1 + c(max(tbl["soma_dim_0"]), max(tbl["soma_dim_1"]))}}

\item{zero_based}{Logical, if TRUE returns a Matrix::\link{sparseMatrix}
otherwise \link{matrixZeroBasedView} of Matrix::\link[Matrix]{sparseMatrix}}
}
\value{
Matrix::\link{sparseMatrix} or \link{matrixZeroBasedView} of Matrix::\link[Matrix]{sparseMatrix}
}
\description{
Converts a \link[arrow]{Table} of sparse format (columns: "soma_dim_0",
"soma_dim_1", "soma_data") to a \link{matrixZeroBasedView}
}
