% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordsStrider.R
\name{CoordsStrider}
\alias{CoordsStrider}
\title{The Coordinate Strider}
\description{
The \code{CoordsStrider} allows creating coordinate slices
in an interated manner. Alternatively, it can chunk an existing vector of
coordinates
}
\note{
The \code{CoordsStrider} operates using
\link[bit64:integer64]{64-bit integer} objects; as such, accessing fields,
such as \code{strider$start} or \code{strider$stride} will return an
\code{integer64} object, which functions differently than a regular
\code{integer}. Use with caution and convert back to integers or numerics
as necessary
}
\examples{
strider <- CoordsStrider$new(start = 1L, end = 200L, stride = 60L)
while (strider$hasNext()) {
  str(strider$nextElem())
}

}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{coords}}{If set, the coordinates to iterate over}

\item{\code{start}}{If set, the starting point of the iterated coordinates;
otherwise the minimum value of \code{self$coords}}

\item{\code{end}}{If set, the end point of the iterated coordinates;
otherwise the maximum value of \code{self$coords}}

\item{\code{stride}}{The stride, or how many coordinates to generate per
iteration; note: this field is settable, which will reset the iterator}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordsStrider-new}{\code{CoordsStrider$new()}}
\item \href{#method-CoordsStrider-print}{\code{CoordsStrider$print()}}
\item \href{#method-CoordsStrider-length}{\code{CoordsStrider$length()}}
\item \href{#method-CoordsStrider-hasNext}{\code{CoordsStrider$hasNext()}}
\item \href{#method-CoordsStrider-nextElem}{\code{CoordsStrider$nextElem()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-new}{}}}
\subsection{Method \code{new()}}{
Create a coordinate strider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordsStrider$new(coords, ..., stride = NULL, start = NULL, end = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coords}}{An integer vector of coordinates}

\item{\code{...}}{Ignored}

\item{\code{stride}}{The stride of how many coordinates to yield per iteration;
by default, will try to yield all coordinates per iteration}

\item{\code{start}}{If \code{coords} is missing, the starting coordinate
to generate}

\item{\code{end}}{If \code{coords} is missing, the ending coordinate
to generate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-print"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-print}{}}}
\subsection{Method \code{print()}}{
Print the coordinate strider to the screen
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordsStrider$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-length"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-length}{}}}
\subsection{Method \code{length()}}{
Get the length (span) of the coordinates within the strider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordsStrider$length()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The length (span) of the coordinate strider
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-hasNext"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-hasNext}{}}}
\subsection{Method \code{hasNext()}}{
Determine if there are more coordinates to yield
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordsStrider$hasNext()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if there are more coordinates to yield or
\code{FALSE} if otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordsStrider-nextElem"></a>}}
\if{latex}{\out{\hypertarget{method-CoordsStrider-nextElem}{}}}
\subsection{Method \code{nextElem()}}{
Generate the next set of coordinates to yield. If there are
no more coordiantes to yield, raises a \code{stopIteration} error
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordsStrider$nextElem()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer vector of the next set of coordinates
}
}
}
