% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_all_metadata}
\alias{get_all_metadata}
\title{Read all metadata (as named list)}
\usage{
get_all_metadata(uri, is_array, ctxxp)
}
\arguments{
\item{uri}{The array URI}

\item{is_array}{A boolean to indicate array or group}

\item{ctxxp}{An external pointer to the SOMAContext wrapper}
}
\description{
This function currently supports metadata as either a string or an 'int64' (or 'int32').
It will error if a different datatype is encountered.
}
