% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAContext.R
\name{get_default_context}
\alias{get_default_context}
\title{Returns the global default context, creating one if necessary}
\usage{
get_default_context()
}
\value{
The context that will be used for TileDB-SOMA API when no context is provided by the user.
}
\description{
Returns the global default context, creating one if necessary
}
