% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_metadata_num}
\alias{get_metadata_num}
\title{Get nnumber of metadata items}
\usage{
get_metadata_num(uri, is_array, ctxxp)
}
\arguments{
\item{uri}{The array URI}

\item{is_array}{A boolean to indicate array or group}

\item{ctxxp}{An external pointer to the SOMAContext wrapper}
}
\description{
Get nnumber of metadata items
}
