% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/get_table.R
\name{get_column_names}
\alias{get_column_names}
\alias{export_column_direct}
\alias{export_arrow_array}
\alias{nnz}
\alias{get_column_types}
\alias{get_table}
\alias{get_column}
\title{Access SOME Data From a Given URI}
\usage{
get_column_names(uri)

export_column_direct(uri, colnames)

export_arrow_array(
  uri,
  colnames,
  qc = NULL,
  dim_points = NULL,
  dim_ranges = NULL,
  loglevel = "warn"
)

get_column_types(uri, colnames)

nnz(uri)

get_table(uri)

get_column(uri, column)
}
\arguments{
\item{uri}{Character value with URI path to a SOMA data set}

\item{colnames}{Vector of character value with the name of the columns to retrieve}

\item{qc}{External Pointer object to TileDB Query Condition, defaults to \sQuote{NULL} i.e.
no query condition}

\item{dim_points}{A data.frame object with name, type and point value to select rows.}

\item{dim_ranges}{A data.frame object with name, type, and low and high values to select
rows by ranges.}

\item{loglevel}{Character value with the desired logging level, defaults to \sQuote{warn}}

\item{column}{Character value with the name of the column to retrieve}
}
\value{
The selected data object(s) from the given data set
}
\description{
These functions access a given SOMA URI and extract, respectively, the
complete data.frame or named column from the data set.
}
\examples{
\dontrun{
uri <- "test/soco/pbmc3k_processed/obs"
column <-  "n_counts"
summary(get_table(uri))
summary(get_columns(uri, column))
columns <- c("n_genes", "louvain")
z <- export_recordbatch(uri, columns)
rb <- arch::from_arch_array(z, arrow::RecordBatch)
z <- export_recordbatch(uri, columns)
tb <- arrow::as_arrow_table(arch::from_arch_array(z, arrow::RecordBatch))
}
}
