% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\alias{get_column}
\alias{arrow_to_dt}
\title{Prototype Interfaces to \code{libtiledbsoma}}
\usage{
get_table(uri)

get_column(uri, column)

arrow_to_dt(arr)
}
\arguments{
\item{uri}{A character variable with the URI of a TileDB (SOMA) Array}

\item{column, colname, colnames}{A character variable with a column name}

\item{arr}{A \code{arch} \code{array} with \code{arrow} representation of a \code{RecordBatch}}

\item{schemaxp}{A external pointer to an \code{arrow} schema object}

\item{arrayxp}{A external pointer to an \code{arrow} array data object}

\item{level}{A character variable with the desired logging level}
}
\value{
The different functions return their respective values
}
\description{
Access functionality of \code{libtiledbsoma}
}
