% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{has_metadata}
\alias{has_metadata}
\title{Check for metadata given key}
\usage{
has_metadata(uri, key, ctxxp)
}
\arguments{
\item{uri}{The array URI}

\item{key}{The array metadata key}

\item{ctxxp}{An external pointer to the SOMAContext wrapper}
}
\description{
Check for metadata given key
}
