% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntIndexer.R
\name{IntIndexer}
\alias{IntIndexer}
\title{The SOMA Re-Indexer}
\description{
A re-indexer for unique integer indices
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IntIndexer-new}{\code{IntIndexer$new()}}
\item \href{#method-IntIndexer-get_indexer}{\code{IntIndexer$get_indexer()}}
\item \href{#method-IntIndexer-clone}{\code{IntIndexer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntIndexer-new"></a>}}
\if{latex}{\out{\hypertarget{method-IntIndexer-new}{}}}
\subsection{Method \code{new()}}{
Create a new re-indexer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntIndexer$new(data, tiledbsoma_ctx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Integer keys used to build the index (hash) table}

\item{\code{tiledbsoma_ctx}}{Optional \code{\link{SOMATileDBContext}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntIndexer-get_indexer"></a>}}
\if{latex}{\out{\hypertarget{method-IntIndexer-get_indexer}{}}}
\subsection{Method \code{get_indexer()}}{
Get the underlying indices for the target data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntIndexer$get_indexer(target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Data to re-index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of 64-bit integers with \code{target} re-indexed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntIndexer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IntIndexer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntIndexer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
