% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrixZeroBasedView.R
\name{matrixZeroBasedView}
\alias{matrixZeroBasedView}
\title{matrixZeroBasedView is a wrapper shim for a matrix or Matrix::sparseMatrix providing}
\description{
\code{matrixZeroBasedView} is a class that allows elemental
matrix access using zero-based indeces.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-matrixZeroBasedView-new}{\code{matrixZeroBasedView$new()}}
\item \href{#method-matrixZeroBasedView-take}{\code{matrixZeroBasedView$take()}}
\item \href{#method-matrixZeroBasedView-dim}{\code{matrixZeroBasedView$dim()}}
\item \href{#method-matrixZeroBasedView-nrow}{\code{matrixZeroBasedView$nrow()}}
\item \href{#method-matrixZeroBasedView-ncol}{\code{matrixZeroBasedView$ncol()}}
\item \href{#method-matrixZeroBasedView-get_one_based_matrix}{\code{matrixZeroBasedView$get_one_based_matrix()}}
\item \href{#method-matrixZeroBasedView-sum}{\code{matrixZeroBasedView$sum()}}
\item \href{#method-matrixZeroBasedView-print}{\code{matrixZeroBasedView$print()}}
\item \href{#method-matrixZeroBasedView-clone}{\code{matrixZeroBasedView$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrixZeroBasedView-new"></a>}}
\if{latex}{\out{\hypertarget{method-matrixZeroBasedView-new}{}}}
\subsection{Method \code{new()}}{
Initialize (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{matrixZeroBasedView$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\link{matrix} or Matrix::\link[Matrix]{sparseMatrix} or Matrix::\link[Matrix]{denseMatrix}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrixZeroBasedView-take"></a>}}
\if{latex}{\out{\hypertarget{method-matrixZeroBasedView-take}{}}}
\subsection{Method \code{take()}}{
Zero-based matrix element access
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{matrixZeroBasedView$take(i = NULL, j = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Row index (zero-based).}

\item{\code{j}}{Column index (zero-based).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The specified matrix slice as another \link{matrixZeroBasedView}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrixZeroBasedView-dim"></a>}}
\if{latex}{\out{\hypertarget{method-matrixZeroBasedView-dim}{}}}
\subsection{Method \code{dim()}}{
dim
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{matrixZeroBasedView$dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dimensions of the matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrixZeroBasedView-nrow"></a>}}
\if{latex}{\out{\hypertarget{method-matrixZeroBasedView-nrow}{}}}
\subsection{Method \code{nrow()}}{
nrow
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{matrixZeroBasedView$nrow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Matrix row count.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrixZeroBasedView-ncol"></a>}}
\if{latex}{\out{\hypertarget{method-matrixZeroBasedView-ncol}{}}}
\subsection{Method \code{ncol()}}{
ncol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{matrixZeroBasedView$ncol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Matrix column count.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrixZeroBasedView-get_one_based_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-matrixZeroBasedView-get_one_based_matrix}{}}}
\subsection{Method \code{get_one_based_matrix()}}{
Get the one-based R matrix with its original class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{matrixZeroBasedView$get_one_based_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
One-based matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrixZeroBasedView-sum"></a>}}
\if{latex}{\out{\hypertarget{method-matrixZeroBasedView-sum}{}}}
\subsection{Method \code{sum()}}{
Perform arithmetic sum between this link{matrixZeroBasedView}
and another link{matrixZeroBasedView}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{matrixZeroBasedView$sum(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the link{matrixZeroBasedView} to sum.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of the sum as a \link{matrixZeroBasedView}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrixZeroBasedView-print"></a>}}
\if{latex}{\out{\hypertarget{method-matrixZeroBasedView-print}{}}}
\subsection{Method \code{print()}}{
print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{matrixZeroBasedView$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matrixZeroBasedView-clone"></a>}}
\if{latex}{\out{\hypertarget{method-matrixZeroBasedView-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{matrixZeroBasedView$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
