% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalarMap.R
\name{ScalarMap}
\alias{ScalarMap}
\title{A Mapping Type for Scalars}
\description{
An R6 mapping type that is limited to scalar atomic vector types only; can
optionally be limited further to a specific atomic vector type
(eg. \dQuote{\code{logical}}).
}
\keyword{internal}
\section{Super class}{
\code{\link[tiledbsoma:MappingBase]{tiledbsoma::MappingBase}} -> \code{ScalarMap}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{The type that this \code{ScalarMap} is limited to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScalarMap-new}{\code{ScalarMap$new()}}
\item \href{#method-ScalarMap-set}{\code{ScalarMap$set()}}
\item \href{#method-ScalarMap-clone}{\code{ScalarMap$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="get"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-get'><code>tiledbsoma::MappingBase$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="items"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-items'><code>tiledbsoma::MappingBase$items()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="keys"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-keys'><code>tiledbsoma::MappingBase$keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="length"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-length'><code>tiledbsoma::MappingBase$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="print"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-print'><code>tiledbsoma::MappingBase$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="remove"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-remove'><code>tiledbsoma::MappingBase$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="setv"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-setv'><code>tiledbsoma::MappingBase$setv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="to_list"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-to_list'><code>tiledbsoma::MappingBase$to_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="update"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-update'><code>tiledbsoma::MappingBase$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="values"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-values'><code>tiledbsoma::MappingBase$values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarMap-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarMap-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScalarMap$new(type = "any")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Limit the \code{ScalarMap} to a preset type; choose from:
NOPENOPENOPE[results=rd]{tiledbsoma:::rd_atomic()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instantiated \code{ScalarMap} object with the
type set to \code{type}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarMap-set"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarMap-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScalarMap$set(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Key to set}

\item{\code{value}}{Value to add for \code{key}, or \code{NULL} to remove
the entry for \code{key}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[chainable] Invisibly returns \code{self} with
\code{value} added as \code{key}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScalarMap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScalarMap-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScalarMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
