% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAContext.R
\name{set_default_context}
\alias{set_default_context}
\title{Create and return a global default context object}
\usage{
set_default_context(config = NULL)
}
\arguments{
\item{config}{...}
}
\value{
The context that will be used for TileDB-SOMA API when no context is provided by the user.
}
\description{
It is recommended to call this method once before all other TileDB-SOMA R API calls. If the global context
was already set, a warning will be raised. Setting a new default context will not change the context for
TileDB-SOMA objects that were already created.
}
