% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tiledb.R
\name{show_package_versions}
\alias{show_package_versions}
\title{Display package versions}
\usage{
show_package_versions()
}
\description{
Print version information for \pkg{tiledb} (R package), libtiledbsoma, and
TileDB embedded, suitable for assisting with bug reports.
}
