% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{soma_context}
\alias{soma_context}
\title{Create and cache a SOMA Context Object}
\usage{
soma_context(config = NULL)
}
\arguments{
\item{config}{A named character vector with \dQuote{key} and \dQuote{value}
pairs defining the configuration setting.}
}
\value{
An external pointer object containing a shared pointer instance
of \code{SOMAContext}.
}
\description{
Create and cache a SOMA Context Object
}
\keyword{internal}
