% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMACollectionBase.R
\name{SOMACollectionBase}
\alias{SOMACollectionBase}
\title{SOMA Collection Base Class}
\value{
SOMA object.
}
\description{
Base class for objects containing persistent collection of SOMA
objects, mapping string keys to any SOMA object.  (lifecycle: experimental)
}
\section{Super classes}{
\code{tiledbsoma::TileDBObject} -> \code{tiledbsoma::TileDBGroup} -> \code{SOMACollectionBase}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{soma_type}}{Retrieve the SOMA object type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMACollectionBase-new}{\code{SOMACollectionBase$new()}}
\item \href{#method-SOMACollectionBase-create}{\code{SOMACollectionBase$create()}}
\item \href{#method-SOMACollectionBase-set}{\code{SOMACollectionBase$set()}}
\item \href{#method-SOMACollectionBase-get}{\code{SOMACollectionBase$get()}}
\item \href{#method-SOMACollectionBase-add_new_collection}{\code{SOMACollectionBase$add_new_collection()}}
\item \href{#method-SOMACollectionBase-add_new_dataframe}{\code{SOMACollectionBase$add_new_dataframe()}}
\item \href{#method-SOMACollectionBase-add_new_dense_ndarray}{\code{SOMACollectionBase$add_new_dense_ndarray()}}
\item \href{#method-SOMACollectionBase-add_new_sparse_ndarray}{\code{SOMACollectionBase$add_new_sparse_ndarray()}}
\item \href{#method-SOMACollectionBase-clone}{\code{SOMACollectionBase$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="class"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-class'><code>tiledbsoma::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="exists"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-exists'><code>tiledbsoma::TileDBObject$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="is_open"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-is_open'><code>tiledbsoma::TileDBObject$is_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="mode"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-mode'><code>tiledbsoma::TileDBObject$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="close"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-close'><code>tiledbsoma::TileDBGroup$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="get_metadata"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-get_metadata'><code>tiledbsoma::TileDBGroup$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="length"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-length'><code>tiledbsoma::TileDBGroup$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="names"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-names'><code>tiledbsoma::TileDBGroup$names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="open"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-open'><code>tiledbsoma::TileDBGroup$open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="print"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-print'><code>tiledbsoma::TileDBGroup$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="remove"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-remove'><code>tiledbsoma::TileDBGroup$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="set_metadata"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-set_metadata'><code>tiledbsoma::TileDBGroup$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="to_data_frame"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-to_data_frame'><code>tiledbsoma::TileDBGroup$to_data_frame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="to_list"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-to_list'><code>tiledbsoma::TileDBGroup$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMACollectionBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMACollectionBase-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SOMACollection}. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollectionBase$new(
  uri,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL,
  internal_use_only = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI of the TileDB group}

\item{\code{platform_config}}{Optional storage-engine specific configuration}

\item{\code{tiledbsoma_ctx}}{optional SOMATileDBContext}

\item{\code{tiledb_timestamp}}{Optional Datetime (POSIXct) for TileDB timestamp}

\item{\code{internal_use_only}}{Character value to signal this is a 'permitted' call,
as \code{new()} is considered internal and should not be called directly.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMACollectionBase-create"></a>}}
\if{latex}{\out{\hypertarget{method-SOMACollectionBase-create}{}}}
\subsection{Method \code{create()}}{
Add a new SOMA object to the collection. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollectionBase$create(internal_use_only = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{internal_use_only}}{Character value to signal this is a 'permitted' call,
as \code{create()} is considered internal and should not be called directly.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMACollectionBase-set"></a>}}
\if{latex}{\out{\hypertarget{method-SOMACollectionBase-set}{}}}
\subsection{Method \code{set()}}{
Add a new SOMA object to the collection. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollectionBase$set(object, name = NULL, relative = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{SOMA object.}

\item{\code{name}}{The name to use for the object. Defaults to the object URI's
base name.}

\item{\code{relative}}{An optional logical value indicating whether the new
object's URI is relative to the collection's URI. If \code{NULL} (the
default), the object's URI is assumed to be relative unless it is a
\verb{tiledb://} URI.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMACollectionBase-get"></a>}}
\if{latex}{\out{\hypertarget{method-SOMACollectionBase-get}{}}}
\subsection{Method \code{get()}}{
Retrieve a SOMA object by name. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollectionBase$get(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the object to retrieve.}

\item{\code{mode}}{Mode to open in}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMACollectionBase-add_new_collection"></a>}}
\if{latex}{\out{\hypertarget{method-SOMACollectionBase-add_new_collection}{}}}
\subsection{Method \code{add_new_collection()}}{
Add a new SOMA collection to this collection. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollectionBase$add_new_collection(object, key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{SOMA collection object.}

\item{\code{key}}{The key to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMACollectionBase-add_new_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-SOMACollectionBase-add_new_dataframe}{}}}
\subsection{Method \code{add_new_dataframe()}}{
Add a new SOMA dataframe to this collection. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollectionBase$add_new_dataframe(
  key,
  schema,
  index_column_names,
  platform_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key to be added.}

\item{\code{schema}}{Arrow schema argument passed on to DataFrame$create()}

\item{\code{index_column_names}}{Index column names passed on to DataFrame$create()}

\item{\code{platform_config}}{A \link[tiledbsoma:PlatformConfig]{platform configuration}
object}

\item{\code{platform_config}}{A \link[tiledbsoma:PlatformConfig]{platform configuration}
object}

\item{\code{platform_config}}{A \link[tiledbsoma:PlatformConfig]{platform configuration}
object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMACollectionBase-add_new_dense_ndarray"></a>}}
\if{latex}{\out{\hypertarget{method-SOMACollectionBase-add_new_dense_ndarray}{}}}
\subsection{Method \code{add_new_dense_ndarray()}}{
Add a new SOMA DenseNdArray to this collection. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollectionBase$add_new_dense_ndarray(
  key,
  type,
  shape,
  platform_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key to be added.}

\item{\code{type}}{an \link[arrow:data-type]{Arrow type} defining the type of each
element in the array.}

\item{\code{shape}}{a vector of integers defining the shape of the array.}

\item{\code{platform_config}}{A \link[tiledbsoma:PlatformConfig]{platform configuration}
object}

\item{\code{platform_config}}{A \link[tiledbsoma:PlatformConfig]{platform configuration}
object}

\item{\code{platform_config}}{A \link[tiledbsoma:PlatformConfig]{platform configuration}
object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMACollectionBase-add_new_sparse_ndarray"></a>}}
\if{latex}{\out{\hypertarget{method-SOMACollectionBase-add_new_sparse_ndarray}{}}}
\subsection{Method \code{add_new_sparse_ndarray()}}{
Add a new SOMA SparseNdArray to this collection. (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollectionBase$add_new_sparse_ndarray(
  key,
  type,
  shape,
  platform_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key to be added.}

\item{\code{type}}{an \link[arrow:data-type]{Arrow type} defining the type of each
element in the array.}

\item{\code{shape}}{a vector of integers defining the shape of the array.}

\item{\code{platform_config}}{A \link[tiledbsoma:PlatformConfig]{platform configuration}
object}

\item{\code{platform_config}}{A \link[tiledbsoma:PlatformConfig]{platform configuration}
object}

\item{\code{platform_config}}{A \link[tiledbsoma:PlatformConfig]{platform configuration}
object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMACollectionBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMACollectionBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollectionBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
