% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMACollectionCreate}
\alias{SOMACollectionCreate}
\title{Create SOMA Collection}
\usage{
SOMACollectionCreate(
  uri,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL
)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{Optional SOMATileDBContext}

\item{tiledb_timestamp}{Optional POSIXct for TileDB timestamp}
}
\description{
Factory function to create a SOMADataFrame for writing, (lifecycle: experimental)
}
