% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMACollectionOpen}
\alias{SOMACollectionOpen}
\title{Open SOMA Collection}
\usage{
SOMACollectionOpen(uri, platform_config = NULL, tiledbsoma_ctx = NULL)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{optional SOMATileDBContext}
}
\description{
Factory function to open a SOMACollection for reading, (lifecycle: experimental)
}
