% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMACollectionOpen}
\alias{SOMACollectionOpen}
\title{Open a SOMA Collection}
\usage{
SOMACollectionOpen(
  uri,
  mode = "READ",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL,
  context = NULL
)
}
\arguments{
\item{uri}{URI for the TileDB object.}

\item{mode}{One of \dQuote{\code{READ}} or \dQuote{\code{WRITE}}.}

\item{platform_config}{Optional platform configuration.}

\item{tiledbsoma_ctx}{Optional (DEPRECATED) SOMATileDBContext.}

\item{tiledb_timestamp}{Optional Datetime (POSIXct) for TileDB timestamp;
defaults to the current time. If not \code{NULL}, all members accessed
through the collection inherit the timestamp.}

\item{context}{Optional \code{SOMAContext} object used for TileDB operations.
If a context is not provided, then the default context will be used. Call
\code{set_default_context} once before other SOMA operations to configure
the default context.}
}
\value{
A \link[tiledbsoma:SOMACollection]{SOMA collection} stored at
\code{uri} opened in mode \code{mode}.
}
\description{
Factory function to open a \link[tiledbsoma:SOMACollection]{SOMA collection}
for reading (lifecycle: maturing).
}
\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
uri <- withr::local_tempfile(pattern = "soma-collection")

(col <- SOMACollectionCreate(uri))
col$add_new_sparse_ndarray("sparse", arrow::float64(), shape = c(100L, 100L))
col$close()

(col <- SOMACollectionOpen(uri))
col$names()

\dontshow{
col$close()
}
\dontshow{\}) # examplesIf}
}
