% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAContext.R
\name{SOMAContext}
\alias{SOMAContext}
\title{SOMA Context}
\description{
Context map for TileDB-SOMA objects
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{handle}}{External pointer to the C++ interface}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMAContext-new}{\code{SOMAContext$new()}}
\item \href{#method-SOMAContext-get_config}{\code{SOMAContext$get_config()}}
\item \href{#method-SOMAContext-get_data_protocol}{\code{SOMAContext$get_data_protocol()}}
\item \href{#method-SOMAContext-is_tiledbv2}{\code{SOMAContext$is_tiledbv2()}}
\item \href{#method-SOMAContext-is_tiledbv3}{\code{SOMAContext$is_tiledbv3()}}
\item \href{#method-SOMAContext-clone}{\code{SOMAContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAContext-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAContext$new(config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instantiated \code{SOMATileDBContext} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAContext-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAContext-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAContext$get_config()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector with the current config options set on the context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAContext-get_data_protocol"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAContext-get_data_protocol}{}}}
\subsection{Method \code{get_data_protocol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAContext$get_data_protocol(uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{A URI for a SOMA object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The data protocol to use for the URI.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAContext-is_tiledbv2"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAContext-is_tiledbv2}{}}}
\subsection{Method \code{is_tiledbv2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAContext$is_tiledbv2(uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{A URI for a SOMA object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the URI will use \code{tiledbv2} semantics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAContext-is_tiledbv3"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAContext-is_tiledbv3}{}}}
\subsection{Method \code{is_tiledbv3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAContext$is_tiledbv3(uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{A URI for a SOMA object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the URI will use \code{tiledbv3} semantics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
