% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMADataFrameOpen}
\alias{SOMADataFrameOpen}
\title{Open SOMA DataFrame}
\usage{
SOMADataFrameOpen(uri, platform_config = NULL, tiledbsoma_ctx = NULL)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{Optional SOMATileDBContext}
}
\description{
Factory function to open a SOMADataFrame for reading, (lifecycle: experimental)
}
