% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMADenseNDArrayOpen}
\alias{SOMADenseNDArrayOpen}
\title{Open SOMA Dense Nd Array}
\usage{
SOMADenseNDArrayOpen(
  uri,
  mode = "READ",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL
)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{mode}{One of \code{"READ"} or \code{"WRITE"}}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{Optional SOMATileDBContext}

\item{tiledb_timestamp}{Optional Datetime (POSIXct) for TileDB timestamp.
In READ mode, defaults to the current time. If non-NULL, then all members
accessed through the collection object inherit the timestamp.}
}
\description{
Factory function to open a SOMADenseNDArray for reading, (lifecycle: maturing)
}
