% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAExperiment.R
\name{SOMAExperiment}
\alias{SOMAExperiment}
\title{SOMA Experiment}
\description{
\code{SOMAExperiment} is a specialized \code{\link{SOMACollection}},
representing one or more modes of measurement across a single collection of
cells (aka a "multimodal dataset") with pre-defined fields: \code{obs} and \code{ms}
(see \emph{Active Bindings} below for details). (lifecycle: experimental)
}
\section{Super classes}{
\code{\link[tiledbsoma:TileDBObject]{tiledbsoma::TileDBObject}} -> \code{\link[tiledbsoma:TileDBGroup]{tiledbsoma::TileDBGroup}} -> \code{\link[tiledbsoma:SOMACollectionBase]{tiledbsoma::SOMACollectionBase}} -> \code{SOMAExperiment}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{obs}}{a \code{\link{SOMADataFrame}} containing primary annotations on the
observation axis. The contents of the \code{soma_joinid} column define the
observation index domain, \code{obs_id}. All observations for the
\code{SOMAExperiment} must be defined in this dataframe.}

\item{\code{ms}}{a \code{\link{SOMACollection}} of named \code{\link{SOMAMeasurement}}s.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMAExperiment-to_seurat}{\code{SOMAExperiment$to_seurat()}}
\item \href{#method-SOMAExperiment-clone}{\code{SOMAExperiment$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="class"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-class'><code>tiledbsoma::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="exists"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-exists'><code>tiledbsoma::TileDBObject$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="is_open"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-is_open'><code>tiledbsoma::TileDBObject$is_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBObject" data-id="mode"><a href='../../tiledbsoma/html/TileDBObject.html#method-TileDBObject-mode'><code>tiledbsoma::TileDBObject$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="close"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-close'><code>tiledbsoma::TileDBGroup$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="get_metadata"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-get_metadata'><code>tiledbsoma::TileDBGroup$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="length"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-length'><code>tiledbsoma::TileDBGroup$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="names"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-names'><code>tiledbsoma::TileDBGroup$names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="open"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-open'><code>tiledbsoma::TileDBGroup$open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="print"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-print'><code>tiledbsoma::TileDBGroup$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="remove"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-remove'><code>tiledbsoma::TileDBGroup$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="set_metadata"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-set_metadata'><code>tiledbsoma::TileDBGroup$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="to_data_frame"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-to_data_frame'><code>tiledbsoma::TileDBGroup$to_data_frame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="TileDBGroup" data-id="to_list"><a href='../../tiledbsoma/html/TileDBGroup.html#method-TileDBGroup-to_list'><code>tiledbsoma::TileDBGroup$to_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="SOMACollectionBase" data-id="add_new_collection"><a href='../../tiledbsoma/html/SOMACollectionBase.html#method-SOMACollectionBase-add_new_collection'><code>tiledbsoma::SOMACollectionBase$add_new_collection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="SOMACollectionBase" data-id="add_new_dataframe"><a href='../../tiledbsoma/html/SOMACollectionBase.html#method-SOMACollectionBase-add_new_dataframe'><code>tiledbsoma::SOMACollectionBase$add_new_dataframe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="SOMACollectionBase" data-id="add_new_dense_ndarray"><a href='../../tiledbsoma/html/SOMACollectionBase.html#method-SOMACollectionBase-add_new_dense_ndarray'><code>tiledbsoma::SOMACollectionBase$add_new_dense_ndarray()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="SOMACollectionBase" data-id="add_new_sparse_ndarray"><a href='../../tiledbsoma/html/SOMACollectionBase.html#method-SOMACollectionBase-add_new_sparse_ndarray'><code>tiledbsoma::SOMACollectionBase$add_new_sparse_ndarray()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="SOMACollectionBase" data-id="create"><a href='../../tiledbsoma/html/SOMACollectionBase.html#method-SOMACollectionBase-create'><code>tiledbsoma::SOMACollectionBase$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="SOMACollectionBase" data-id="get"><a href='../../tiledbsoma/html/SOMACollectionBase.html#method-SOMACollectionBase-get'><code>tiledbsoma::SOMACollectionBase$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="SOMACollectionBase" data-id="initialize"><a href='../../tiledbsoma/html/SOMACollectionBase.html#method-SOMACollectionBase-initialize'><code>tiledbsoma::SOMACollectionBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="SOMACollectionBase" data-id="set"><a href='../../tiledbsoma/html/SOMACollectionBase.html#method-SOMACollectionBase-set'><code>tiledbsoma::SOMACollectionBase$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperiment-to_seurat"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperiment-to_seurat}{}}}
\subsection{Method \code{to_seurat()}}{
Loads the experiment as a \code{\link[SeuratObject]{Seurat}}
object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperiment$to_seurat(
  X_layers,
  obs_index = NULL,
  var_index = NULL,
  obs_column_names = NULL,
  var_column_names = NULL,
  obsm_layers = NULL,
  varm_layers = NULL,
  obsp_layers = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_layers}}{A named list of named character vectors describing the
measurements to load and the layers within those measurements to read in;
for example: \preformatted{
list(
  RNA = c(counts = "counts", data = "logcounts"),
  ADT = c(counts = "counts")
)
}}

\item{\code{obs_index}}{Name of column in \code{obs} to add as cell names; uses
\code{paste0("cell", obs_joinids())} by default}

\item{\code{var_index}}{A named character of column names in \code{var} for
each measurement to use as feature names; for example: \preformatted{
c(RNA = "gene_name", ADT = "protein_name")
}
Uses \code{paste0("feature", var_joinids())} by default}

\item{\code{obs_column_names}}{Names of columns in \code{obs} to add as
cell-level meta data; by default, loads all columns}

\item{\code{var_column_names}}{A named list of character vectors describing the
columns in \code{var} for each measurement to load in as feature-level
meta data; for example: \preformatted{
list(
  RNA = c("vst.mean", "vst.variable"),
  ADT = c("ensembl_id")
)
}
By default, loads in entire feature-level meta data for all measurements
described in \code{X_layers}}

\item{\code{obsm_layers}}{A named list of character vectors describing the
arrays in \code{obsm} for each measurement to load in as
dimensional reductions; for example: \preformatted{
list(
  RNA = c("pca", "umap"),
  ADT = c("adtpca", "adtumap")
)
}
By default, loads in all dimensional reductions for all measurements
described in \code{X_layers}}

\item{\code{varm_layers}}{A named list of named character vectors describing the
arrays in \code{varm} to load in as feature loadings and which array in
\code{obsm} they correspond to; for example: \preformatted{
list(
  RNA = c(pca = "PCs"),
  ADT = c(adtpca = "ADTPCs")
)
}
By default, will try to determine \code{varm_layers} from
\code{obsm_layers} and load in all loadings for all dimensional
reductions for all measurements described in \code{X_layers}}

\item{\code{obsp_layers}}{A named list of character vectors describing the
arrays in \code{obsp} for each measurement to load in as
nearest neighbor graphs; for example: \preformatted{
list(
  RNA = c("RNA_nn", "RNA_snn"),
  ADT = c("ADT_nn")
)
}
By default, loads in all nearest-neighbor graphs for all measurements for
all measurements described in \code{X_layers}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[SeuratObject]{Seurat}} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAExperiment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAExperiment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAExperiment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
