% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMAMeasurementOpen}
\alias{SOMAMeasurementOpen}
\title{Open SOMA Measurement}
\usage{
SOMAMeasurementOpen(uri, platform_config = NULL, tiledbsoma_ctx = NULL)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{optional SOMATileDBContext}
}
\description{
Factory function to open a SOMAMeasurement for reading, (lifecycle: experimental)
}
