% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMAMeasurementOpen}
\alias{SOMAMeasurementOpen}
\title{Open SOMA Measurement}
\usage{
SOMAMeasurementOpen(
  uri,
  mode = "READ",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL,
  context = NULL
)
}
\arguments{
\item{uri}{URI for the TileDB object.}

\item{mode}{One of \dQuote{\code{READ}} or \dQuote{\code{WRITE}}.}

\item{platform_config}{Optional platform configuration.}

\item{tiledbsoma_ctx}{Optional (DEPRECATED) SOMATileDBContext.}

\item{tiledb_timestamp}{Optional Datetime (POSIXct) for TileDB timestamp;
defaults to the current time. If not \code{NULL}, all members accessed
through the collection inherit the timestamp.}

\item{context}{Optional \code{SOMAContext} object used for TileDB operations.
If a context is not provided, then the default context will be used. Call
\code{set_default_context} once before other SOMA operations to configure
the default context.}
}
\value{
A \link[tiledbsoma:SOMAMeasurement]{SOMA measurement} stored at
\code{uri} opened in mode \code{mode}.
}
\description{
Factory function to open a
\link[tiledbsoma:SOMAMeasurement]{SOMA measurement} for reading
(lifecycle: maturing).
}
\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
uri <- withr::local_tempfile(pattern = "soma-measurement")
var <- data.frame(
  soma_joinid = bit64::seq.integer64(0L, 99L),
  var_id = paste0("feature_", seq_len(100L))
)
sch <- arrow::infer_schema(var)

(ms <- SOMAMeasurementCreate(uri))
sdf <- ms$add_new_dataframe(
  "var",
  sch,
  "soma_joinid",
  list(soma_joinid = c(0, 100))
)
sdf$write(arrow::as_arrow_table(var, schema = sch))
sdf$close()
ms$close()

(ms <- SOMAMeasurementOpen(uri))
ms$var

\dontshow{
ms$close()
}
\dontshow{\}) # examplesIf}
}
