% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAObject.R
\name{SOMAObject}
\alias{SOMAObject}
\title{The SOMA Object Base Class}
\description{
Base class to implement shared functionality across the
\code{\link{SOMAArrayBase}} and \code{\link{SOMACollectionBase}} classes.
(lifecycle: maturing)
}
\seealso{
Derived classes: \code{\link{SOMAArrayBase}},
\code{\link{SOMACollectionBase}}
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{platform_config}}{Platform configuration}

\item{\code{context}}{SOMAContext context object for TileDB operations}

\item{\code{tiledbsoma_ctx}}{SOMATileDBContext}

\item{\code{tiledb_timestamp}}{Time that object was opened at}

\item{\code{uri}}{The URI of the TileDB object}

\item{\code{soma_type}}{The SOMA object type}

\item{\code{.tiledb_timestamp_range}}{Time range for libtiledbsoma}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMAObject-new}{\code{SOMAObject$new()}}
\item \href{#method-SOMAObject-is_open}{\code{SOMAObject$is_open()}}
\item \href{#method-SOMAObject-class}{\code{SOMAObject$class()}}
\item \href{#method-SOMAObject-mode}{\code{SOMAObject$mode()}}
\item \href{#method-SOMAObject-reopen}{\code{SOMAObject$reopen()}}
\item \href{#method-SOMAObject-exists}{\code{SOMAObject$exists()}}
\item \href{#method-SOMAObject-get_metadata}{\code{SOMAObject$get_metadata()}}
\item \href{#method-SOMAObject-set_metadata}{\code{SOMAObject$set_metadata()}}
\item \href{#method-SOMAObject-print}{\code{SOMAObject$print()}}
\item \href{#method-SOMAObject-clone}{\code{SOMAObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-new}{}}}
\subsection{Method \code{new()}}{
Create a new SOMA object. (lifecycle: maturing)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$new(
  uri,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL,
  context = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI for the SOMA object}

\item{\code{...}}{Ignored}

\item{\code{platform_config}}{Optional platform configuration}

\item{\code{tiledbsoma_ctx}}{Optional (DEPRECATED) TileDB SOMA context}

\item{\code{tiledb_timestamp}}{Optional timestamp (\code{\link[base]{POSIXct}})
to open the object at}

\item{\code{context}}{A \code{SOMAContext} object. Required internally but
automatically provided by factory functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-is_open"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-is_open}{}}}
\subsection{Method \code{is_open()}}{
Determine if the object is open for reading or writing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$is_open()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the object is open, otherwise \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-class"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-class}{}}}
\subsection{Method \code{class()}}{
Print the name of the R6 class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$class()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The name of the R6 class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-mode"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-mode}{}}}
\subsection{Method \code{mode()}}{
Get the mode of the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$mode()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The mode of the object, one of:
\itemize{
\item \dQuote{\code{CLOSED}}
\item \dQuote{\code{READ}}
\item \dQuote{\code{WRITE}}
\item \dQuote{\code{DELETE}}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-reopen"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-reopen}{}}}
\subsection{Method \code{reopen()}}{
Close and reopen the TileDB object in a new mode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$reopen(mode, tiledb_timestamp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{New mode to open the object in; choose from:
\itemize{
\item \dQuote{\code{READ}}
\item \dQuote{\code{WRITE}}
\item \dQuote{\code{DELETE}}
}}

\item{\code{tiledb_timestamp}}{Optional Datetime (POSIXct) with TileDB timestamp}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns \code{self} opened in \code{mode}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-exists"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-exists}{}}}
\subsection{Method \code{exists()}}{
Check if the object exists. (lifecycle: maturing)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$exists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the object exists, otherwise \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Retrieve metadata. (lifecycle: maturing)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$get_metadata(key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The name of the metadata attribute to retrieve
is not NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of metadata values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-set_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-set_metadata}{}}}
\subsection{Method \code{set_metadata()}}{
Add list of metadata. (lifecycle: maturing)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$set_metadata(metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Named list of metadata to add}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns \code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-print}{}}}
\subsection{Method \code{print()}}{
Print-friendly representation of the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisibly returns \code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMAObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMAObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMAObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
