% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAOpen.R
\name{SOMAOpen}
\alias{SOMAOpen}
\title{Open a SOMA Object}
\usage{
SOMAOpen(
  uri,
  mode = "READ",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  tiledb_timestamp = NULL
)
}
\arguments{
\item{uri}{URI for the TileDB object.}

\item{mode}{One of \dQuote{\code{READ}} or \dQuote{\code{WRITE}}}

\item{platform_config}{Optional platform configuration.}

\item{tiledbsoma_ctx}{Optional (DEPRECATED) SOMATileDBContext.}

\item{context}{Optional \code{SOMAContext} object used for TileDB operations.
If a context is not provided, then the default context will be used. Call
\code{set_default_context} once before other SOMA operations to configure
the default context.}

\item{tiledb_timestamp}{Optional Datetime (POSIXct) for TileDB timestamp;
defaults to the current time. If not \code{NULL}, all members accessed
through the collection inherit the timestamp.}
}
\value{
A SOMA object
}
\description{
Utility function to open the corresponding SOMA object given a URI
(lifecycle: maturing).
}
\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dir <- withr::local_tempfile(pattern = "soma-open")
dir.create(dir, recursive = TRUE)

uri <- extract_dataset("soma-exp-pbmc-small", dir)
(exp <- SOMAOpen(uri))

\dontshow{
exp$close()
}

uri <- extract_dataset("soma-dataframe-pbmc3k-processed-obs", dir)
(obs <- SOMAOpen(uri))

\dontshow{
obs$close()
}
\dontshow{\}) # examplesIf}
}
