% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMAOpen.R
\name{SOMAOpen}
\alias{SOMAOpen}
\title{Open a SOMA Object}
\usage{
SOMAOpen(
  uri,
  mode = "READ",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL
)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{mode}{One of \code{"READ"} or \code{"WRITE"}}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{Optional SOMATileDBContext}

\item{tiledb_timestamp}{Optional Datetime (POSIXct) with TileDB timestamp}
}
\description{
Utility function to open the corresponding SOMA Object given a URI, (lifecycle: experimental)
}
