% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMASparseNDArrayRead.R
\name{SOMASparseNDArrayBlockwiseRead}
\alias{SOMASparseNDArrayBlockwiseRead}
\title{Blockwise Sparse ND-Array Reader}
\description{
Blockwise reader for \code{\link{SOMASparseNDArray}}
}
\keyword{internal}
\section{Super class}{
\code{\link[tiledbsoma:SOMASparseNDArrayReadBase]{tiledbsoma::SOMASparseNDArrayReadBase}} -> \code{SOMASparseNDArrayBlockwiseRead}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{axis}}{The axis to iterate over in a blockwise fashion}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMASparseNDArrayBlockwiseRead-new}{\code{SOMASparseNDArrayBlockwiseRead$new()}}
\item \href{#method-SOMASparseNDArrayBlockwiseRead-tables}{\code{SOMASparseNDArrayBlockwiseRead$tables()}}
\item \href{#method-SOMASparseNDArrayBlockwiseRead-sparse_matrix}{\code{SOMASparseNDArrayBlockwiseRead$sparse_matrix()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayBlockwiseRead-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayBlockwiseRead-new}{}}}
\subsection{Method \code{new()}}{
Create
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayBlockwiseRead$new(
  sr,
  array,
  coords,
  axis,
  ...,
  size,
  reindex_disable_on_axis = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sr}}{SOMA read pointer}

\item{\code{array}}{Underlying \code{\link{SOMASparseNDArray}}}

\item{\code{coords}}{Optional named list of \code{\link[bit64]{integer64}}
values; must be named after \code{array$dimnames()}}

\item{\code{axis}}{Axis to iterate over in a blockwise manner}

\item{\code{...}}{Ignored}

\item{\code{size}}{The size of each blockwise chunk to generate}

\item{\code{reindex_disable_on_axis}}{Additional axes that will not be re-indexed;
currently not yet implemented}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayBlockwiseRead-tables"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayBlockwiseRead-tables}{}}}
\subsection{Method \code{tables()}}{
Read as an \code{\link[Arrow:Table]{Arrow::Table}}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayBlockwiseRead$tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A blockwise iterator yielding chunks as
\code{\link[Arrow:Table]{Arrow::Table}s}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayBlockwiseRead-sparse_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayBlockwiseRead-sparse_matrix}{}}}
\subsection{Method \code{sparse_matrix()}}{
Read as a sparse matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayBlockwiseRead$sparse_matrix(repr = "T")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repr}}{Representation of the sparse matrix to return; choose from:
\itemize{
\item \dQuote{\code{T}}: returns a
\code{\link[Matrix:TsparseMatrix-class]{TsparseMatrix}}
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A blockwise iterator yielding chunks as sparse matrices
}
}
}
