% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factory.R
\name{SOMASparseNDArrayCreate}
\alias{SOMASparseNDArrayCreate}
\title{Create SOMA Sparse Nd Array}
\usage{
SOMASparseNDArrayCreate(
  uri,
  type,
  shape,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL
)
}
\arguments{
\item{uri}{URI for the TileDB object}

\item{type}{An \link[arrow:data-type]{Arrow type} defining the type of each element in the array.}

\item{shape}{A vector of integers defining the shape of the array.}

\item{platform_config}{Optional platform configuration}

\item{tiledbsoma_ctx}{Optional SOMATileDBContext}

\item{tiledb_timestamp}{Optional Datetime (POSIXct) for TileDB timestamp}
}
\description{
Factory function to create a SOMASparseNDArray for writing, (lifecycle: experimental)
}
