% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMASparseNDArrayRead.R
\name{SOMASparseNDArrayRead}
\alias{SOMASparseNDArrayRead}
\title{SOMASparseNDArrayRead}
\description{
Intermediate type to choose result format when reading a sparse array
}
\keyword{internal}
\section{Super class}{
\code{\link[tiledbsoma:SOMASparseNDArrayReadBase]{tiledbsoma::SOMASparseNDArrayReadBase}} -> \code{SOMASparseNDArrayRead}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMASparseNDArrayRead-sparse_matrix}{\code{SOMASparseNDArrayRead$sparse_matrix()}}
\item \href{#method-SOMASparseNDArrayRead-tables}{\code{SOMASparseNDArrayRead$tables()}}
\item \href{#method-SOMASparseNDArrayRead-blockwise}{\code{SOMASparseNDArrayRead$blockwise()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="SOMASparseNDArrayReadBase" data-id="initialize"><a href='../../tiledbsoma/html/SOMASparseNDArrayReadBase.html#method-SOMASparseNDArrayReadBase-initialize'><code>tiledbsoma::SOMASparseNDArrayReadBase$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayRead-sparse_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayRead-sparse_matrix}{}}}
\subsection{Method \code{sparse_matrix()}}{
Read as a sparse matrix (lifecycle: maturing). Returns
an iterator of Matrix::\link[Matrix]{dgTMatrix-class} or
\link{matrixZeroBasedView} of it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayRead$sparse_matrix(zero_based = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zero_based}}{Logical, if \code{TRUE} returns iterator of
\link{matrixZeroBasedView} if \code{FALSE} returns iterator of
Matrix::\link[Matrix]{dgTMatrix-class}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{SparseReadIter}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayRead-tables"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayRead-tables}{}}}
\subsection{Method \code{tables()}}{
Read as a arrow::\link[arrow]{Table} (lifecycle: maturing).
Returns an iterator of arrow::\link[arrow]{Table}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayRead$tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{TableReadIter}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMASparseNDArrayRead-blockwise"></a>}}
\if{latex}{\out{\hypertarget{method-SOMASparseNDArrayRead-blockwise}{}}}
\subsection{Method \code{blockwise()}}{
Read in a blockwise fashion
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMASparseNDArrayRead$blockwise(
  axis,
  ...,
  size = NULL,
  reindex_disable_on_axis = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{axis}}{Axis to iterate over in a blockwise manner}

\item{\code{...}}{Ignored}

\item{\code{size}}{The size of each blockwise chunk to generate}

\item{\code{reindex_disable_on_axis}}{Additional axes that will not be re-indexed;
the following values may be used as shorthands for common settings:
\itemize{
\item \dQuote{\code{TRUE}}: disable re-indexing on all axes
\item \dQuote{\code{NA}}: re-index only on \code{axis}, disable
re-indexing on all others
\item \dQuote{\code{FALSE}}: re-index on \emph{all} axes, do \strong{not}
disable re-indexing
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{SOMASparseNDArrayBlockwiseRead}} iterated reader
}
}
}
