% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMATileDBContext.R
\name{SOMATileDBContext}
\alias{SOMATileDBContext}
\title{SOMA TileDB Context}
\description{
Context map for TileDB-backed SOMA objects
}
\section{Super classes}{
\code{tiledbsoma::MappingBase} -> \code{tiledbsoma::ScalarMap} -> \code{tiledbsoma::SOMAContextBase} -> \code{SOMATileDBContext}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SOMATileDBContext-new}{\code{SOMATileDBContext$new()}}
\item \href{#method-SOMATileDBContext-keys}{\code{SOMATileDBContext$keys()}}
\item \href{#method-SOMATileDBContext-items}{\code{SOMATileDBContext$items()}}
\item \href{#method-SOMATileDBContext-length}{\code{SOMATileDBContext$length()}}
\item \href{#method-SOMATileDBContext-get}{\code{SOMATileDBContext$get()}}
\item \href{#method-SOMATileDBContext-set}{\code{SOMATileDBContext$set()}}
\item \href{#method-SOMATileDBContext-to_tiledb_context}{\code{SOMATileDBContext$to_tiledb_context()}}
\item \href{#method-SOMATileDBContext-context}{\code{SOMATileDBContext$context()}}
\item \href{#method-SOMATileDBContext-clone}{\code{SOMATileDBContext$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="print"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-print'><code>tiledbsoma::MappingBase$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="remove"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-remove'><code>tiledbsoma::MappingBase$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="setv"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-setv'><code>tiledbsoma::MappingBase$setv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="to_list"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-to_list'><code>tiledbsoma::MappingBase$to_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="update"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-update'><code>tiledbsoma::MappingBase$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="MappingBase" data-id="values"><a href='../../tiledbsoma/html/MappingBase.html#method-MappingBase-values'><code>tiledbsoma::MappingBase$values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMATileDBContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-SOMATileDBContext-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMATileDBContext$new(config = NULL, cached = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{...}

\item{\code{cached}}{Force new creation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instantiated \code{SOMATileDBContext} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMATileDBContext-keys"></a>}}
\if{latex}{\out{\hypertarget{method-SOMATileDBContext-keys}{}}}
\subsection{Method \code{keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMATileDBContext$keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The keys of the map
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMATileDBContext-items"></a>}}
\if{latex}{\out{\hypertarget{method-SOMATileDBContext-items}{}}}
\subsection{Method \code{items()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMATileDBContext$items()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Return the items of the map as a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMATileDBContext-length"></a>}}
\if{latex}{\out{\hypertarget{method-SOMATileDBContext-length}{}}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMATileDBContext$length()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of items in the map
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMATileDBContext-get"></a>}}
\if{latex}{\out{\hypertarget{method-SOMATileDBContext-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMATileDBContext$get(key, default = quote(expr = ))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Key to fetch}

\item{\code{default}}{Default value to fetch if \code{key} is not found;
defaults to \code{NULL}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The value of \code{key} in the map, or \code{default} if
\code{key} is not found
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMATileDBContext-set"></a>}}
\if{latex}{\out{\hypertarget{method-SOMATileDBContext-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMATileDBContext$set(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Key to set}

\item{\code{value}}{Value to add for \code{key}, or \code{NULL} to remove
the entry for \code{key}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\[chainable\] Invisibly returns \code{self} with
\code{value} added as \code{key}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMATileDBContext-to_tiledb_context"></a>}}
\if{latex}{\out{\hypertarget{method-SOMATileDBContext-to_tiledb_context}{}}}
\subsection{Method \code{to_tiledb_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMATileDBContext$to_tiledb_context()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[tiledb:tiledb_ctx]{tiledb_ctx}} object, dynamically
constructed. Most useful for the constructor of this class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMATileDBContext-context"></a>}}
\if{latex}{\out{\hypertarget{method-SOMATileDBContext-context}{}}}
\subsection{Method \code{context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMATileDBContext$context()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[tiledb:tiledb_ctx]{tiledb_ctx}} object, which is
a stored (and long-lived) result from \code{to_tiledb_context}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SOMATileDBContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SOMATileDBContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMATileDBContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
