% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseReadIter.R
\name{SparseReadIter}
\alias{SparseReadIter}
\title{SparseReadIter}
\description{
\code{SparseReadIter} is a class that allows for iteration over
a reads on \link{SOMASparseNDArray}.
Iteration chunks are retrieved as 0-based Views \link{matrixZeroBasedView} of Matrix::\link[Matrix]{SparseMatrix}.
}
\section{Super class}{
\code{\link[tiledbsoma:ReadIter]{tiledbsoma::ReadIter}} -> \code{SparseReadIter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SparseReadIter-new}{\code{SparseReadIter$new()}}
\item \href{#method-SparseReadIter-concat}{\code{SparseReadIter$concat()}}
\item \href{#method-SparseReadIter-clone}{\code{SparseReadIter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="read_complete"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-read_complete'><code>tiledbsoma::ReadIter$read_complete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsoma" data-topic="ReadIter" data-id="read_next"><a href='../../tiledbsoma/html/ReadIter.html#method-ReadIter-read_next'><code>tiledbsoma::ReadIter$read_next()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SparseReadIter-new"></a>}}
\if{latex}{\out{\hypertarget{method-SparseReadIter-new}{}}}
\subsection{Method \code{new()}}{
Create (lifecycle: experimental)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SparseReadIter$new(sr, shape, zero_based = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sr}}{Soma reader pointer}

\item{\code{shape}}{Shape of the full matrix}

\item{\code{zero_based}}{Logical, if TRUE will make iterator for Matrix::\link[Matrix]{dgTMatrix}
otherwise \link{matrixZeroBasedView}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SparseReadIter-concat"></a>}}
\if{latex}{\out{\hypertarget{method-SparseReadIter-concat}{}}}
\subsection{Method \code{concat()}}{
Concatenate remainder of iterator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SparseReadIter$concat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{matrixZeroBasedView} of Matrix::\link[Matrix]{SparseMatrix}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SparseReadIter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SparseReadIter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SparseReadIter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
