% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBGroup.R
\name{TileDBGroup}
\alias{TileDBGroup}
\title{TileDB Group Base Class}
\value{
A named list of group members.

The array object.
}
\description{
Base class for interacting with TileDB groups
}
\details{
\subsection{Initialization}{

Upon initialization a new group is created if one does not already exist at
the specified \code{uri}.
}
}
\section{Super class}{
\code{\link[tiledbsc:TileDBObject]{tiledbsc::TileDBObject}} -> \code{TileDBGroup}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{members}}{Named list of members in the group}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TileDBGroup-new}{\code{TileDBGroup$new()}}
\item \href{#method-TileDBGroup-print}{\code{TileDBGroup$print()}}
\item \href{#method-TileDBGroup-group_exists}{\code{TileDBGroup$group_exists()}}
\item \href{#method-TileDBGroup-tiledb_group}{\code{TileDBGroup$tiledb_group()}}
\item \href{#method-TileDBGroup-list_objects}{\code{TileDBGroup$list_objects()}}
\item \href{#method-TileDBGroup-list_object_uris}{\code{TileDBGroup$list_object_uris()}}
\item \href{#method-TileDBGroup-add_member}{\code{TileDBGroup$add_member()}}
\item \href{#method-TileDBGroup-remove_member}{\code{TileDBGroup$remove_member()}}
\item \href{#method-TileDBGroup-count_members}{\code{TileDBGroup$count_members()}}
\item \href{#method-TileDBGroup-list_members}{\code{TileDBGroup$list_members()}}
\item \href{#method-TileDBGroup-list_member_uris}{\code{TileDBGroup$list_member_uris()}}
\item \href{#method-TileDBGroup-get_members}{\code{TileDBGroup$get_members()}}
\item \href{#method-TileDBGroup-get_member}{\code{TileDBGroup$get_member()}}
\item \href{#method-TileDBGroup-get_metadata}{\code{TileDBGroup$get_metadata()}}
\item \href{#method-TileDBGroup-add_metadata}{\code{TileDBGroup$add_metadata()}}
\item \href{#method-TileDBGroup-clone}{\code{TileDBGroup$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBObject" data-id="class"><a href='../../tiledbsc/html/TileDBObject.html#method-TileDBObject-class'><code>tiledbsc::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBObject" data-id="exists"><a href='../../tiledbsc/html/TileDBObject.html#method-TileDBObject-exists'><code>tiledbsc::TileDBObject$exists()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-new}{}}}
\subsection{Method \code{new()}}{
Create a new TileDBGroup object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$new(uri, verbose = TRUE, config = NULL, ctx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{TileDB array URI}

\item{\code{verbose}}{Print status messages}

\item{\code{config}}{optional configuration}

\item{\code{ctx}}{optional tiledb context}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-print"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-print}{}}}
\subsection{Method \code{print()}}{
Print summary of the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-group_exists"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-group_exists}{}}}
\subsection{Method \code{group_exists()}}{
Check if the group exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$group_exists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if the group exists, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-tiledb_group"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-tiledb_group}{}}}
\subsection{Method \code{tiledb_group()}}{
Return a \code{\link{tiledb_group}} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$tiledb_group(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments to pass to \code{tiledb::tiledb_array()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[tiledb:tiledb_group]{tiledb::tiledb_group}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-list_objects"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-list_objects}{}}}
\subsection{Method \code{list_objects()}}{
List the TileDB objects within the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_objects(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of object to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all object types are listed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with columns \code{URI} and \code{TYPE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-list_object_uris"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-list_object_uris}{}}}
\subsection{Method \code{list_object_uris()}}{
List URIs for TileDB objects within the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_object_uris(type = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of object to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all object types are listed.}

\item{\code{prefix}}{Filter URIs whose basename contain an optional prefix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of object URIs with names corresponding to the
basename of the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-add_member"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-add_member}{}}}
\subsection{Method \code{add_member()}}{
Add new member to the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$add_member(object, name = NULL, relative = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{The \code{TileDBArray} or \code{TileDBGroup} object to add.}

\item{\code{name}}{The name to use for the member. By default the base name of
the object's URI is used.}

\item{\code{relative}}{A logical value indicating whether the new member's URI
is relative to the group's URI.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-remove_member"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-remove_member}{}}}
\subsection{Method \code{remove_member()}}{
Remove member from the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$remove_member(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the member to remove.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-count_members"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-count_members}{}}}
\subsection{Method \code{count_members()}}{
Count the number of members in the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$count_members()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer count of members in the group.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-list_members"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-list_members}{}}}
\subsection{Method \code{list_members()}}{
List the members of the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_members(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of member to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all member types are listed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with columns \code{URI}, \code{TYPE}, and \code{NAME}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-list_member_uris"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-list_member_uris}{}}}
\subsection{Method \code{list_member_uris()}}{
List URIs for group members
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_member_uris(type = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of member to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all member types are listed.}

\item{\code{prefix}}{Filter for members whose name contains an optional prefix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of member URIs, named for the group member
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-get_members"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-get_members}{}}}
\subsection{Method \code{get_members()}}{
Retrieve arrays within the group that meet the specified
criteria.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_members(type = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of group members to list, either \code{"ARRAY"}, or
\code{"GROUP"}.}

\item{\code{prefix}}{String prefix to filter the member names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-get_member"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-get_member}{}}}
\subsection{Method \code{get_member()}}{
Retrieve a group member.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_member(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the array to retrieve.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Retrieve metadata from the TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_metadata(key = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The name of the metadata attribute to retrieve.}

\item{\code{prefix}}{Filter metadata using an optional prefix. Ignored if \code{key}
is not NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of metadata values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-add_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-add_metadata}{}}}
\subsection{Method \code{add_metadata()}}{
Add list of metadata to the TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$add_metadata(metadata, prefix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Named list of metadata to add.}

\item{\code{prefix}}{Optional prefix to add to the metadata attribute names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
