% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBObject.R
\name{TileDBObject}
\alias{TileDBObject}
\title{TileDB Object Base Class}
\description{
Base class to implement shared functionality across the TileDBArray and
TileDBGroup classes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{verbose}}{Whether to print verbose output}

\item{\code{config}}{optional configuration}

\item{\code{ctx}}{optional tiledb context}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{uri}}{The URI of the TileDB object.}

\item{\code{object}}{Access the underlying TileB object directly (either a
\code{\link[tiledb:tiledb_array]{tiledb::tiledb_array}} or \code{\link[tiledb:tiledb_group]{tiledb::tiledb_group}}).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TileDBObject-new}{\code{TileDBObject$new()}}
\item \href{#method-TileDBObject-class}{\code{TileDBObject$class()}}
\item \href{#method-TileDBObject-print}{\code{TileDBObject$print()}}
\item \href{#method-TileDBObject-exists}{\code{TileDBObject$exists()}}
\item \href{#method-TileDBObject-clone}{\code{TileDBObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-new}{}}}
\subsection{Method \code{new()}}{
Create a new TileDB object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$new(uri, verbose = TRUE, config = NULL, ctx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI for the TileDB object}

\item{\code{verbose}}{Print status messages}

\item{\code{config}}{optional configuration}

\item{\code{ctx}}{optional TileDB context}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-class"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-class}{}}}
\subsection{Method \code{class()}}{
Print the name of the R6 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-print}{}}}
\subsection{Method \code{print()}}{
Print summary of the array.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-exists"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-exists}{}}}
\subsection{Method \code{exists()}}{
Check if the object exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$exists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\verb{TRUE`` if the object exists, }FALSE` otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
