% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utils-tiledb.R
\name{tiledbsoma_stats}
\alias{tiledbsoma_stats}
\alias{tiledbsoma_stats_enable}
\alias{tiledbsoma_stats_disable}
\alias{tiledbsoma_stats_reset}
\alias{tiledbsoma_stats_dump}
\alias{tiledbsoma_stats_show}
\title{TileDB SOMA statistics}
\usage{
tiledbsoma_stats_enable()

tiledbsoma_stats_disable()

tiledbsoma_stats_reset()

tiledbsoma_stats_dump()

tiledbsoma_stats_show()
}
\description{
These functions expose the TileDB Core functionality for performance measurements
and statistics.
}
\details{
\itemize{
\item \code{tiledbsoma_stats_enable()}/\code{tiledbsoma_stats_disable()}: Enable and disable TileDB's internal statistics.
\item \code{tiledbsoma_stats_reset()}: Reset all statistics to 0.
\item \code{tiledbsoma_stats_dump()}: Dump all statistics to a JSON string.
\item \code{tiledbsoma_stats_show()}: Print all statistics to the console.
}
}
