% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_soma.R
\name{write_soma}
\alias{write_soma}
\title{Write a SOMA Object from an \R Object}
\usage{
write_soma(
  x,
  uri,
  ...,
  mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL
)
}
\arguments{
\item{x}{An object}

\item{uri}{URI for resulting SOMA object}

\item{...}{Arguments passed to other methods}

\item{mode}{Ingestion mode: one of \code{write} or \code{resume}}

\item{platform_config}{Optional \link[tiledbsoma:PlatformConfig]{platform
configuration}}

\item{tiledbsoma_ctx}{Optional \code{\link{SOMATileDBContext}}}
}
\value{
The URI to the resulting \code{\link{SOMAExperiment}} generated from
the data contained in \code{x}
}
\description{
Convert \R objects to their appropriate SOMA counterpart
function and methods can be written for it to provide a high-level
\R \eqn{\rightarrow} SOMA interface
}
\section{Known methods}{

\itemize{
\item \link[tiledbsoma:write_soma.Seurat]{Writing Seurat objects}
\item \link[tiledbsoma:write_soma.SummarizedExperiment]{Writing SummarizedExperiment objects}
\item \link[tiledbsoma:write_soma.SingleCellExperiment]{Writing SingleCellExperiment objects}
}
}

