% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bioc.R, R/write_soma.R
\name{write_soma.DataFrame}
\alias{write_soma.DataFrame}
\alias{write_soma.Hits}
\alias{write_soma_objects}
\alias{write_soma.character}
\alias{write_soma.data.frame}
\alias{write_soma.matrix}
\alias{write_soma.Matrix}
\alias{write_soma.TsparseMatrix}
\title{Write R Objects to SOMA}
\usage{
\method{write_soma}{DataFrame}(
  x,
  uri,
  soma_parent,
  df_index = NULL,
  index_column_names = "soma_joinid",
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

\method{write_soma}{Hits}(
  x,
  uri,
  soma_parent,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

\method{write_soma}{character}(
  x,
  uri,
  soma_parent,
  ...,
  key = NULL,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

\method{write_soma}{data.frame}(
  x,
  uri,
  soma_parent,
  df_index = NULL,
  index_column_names = "soma_joinid",
  ...,
  key = NULL,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

\method{write_soma}{matrix}(
  x,
  uri,
  soma_parent,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  key = NULL,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

\method{write_soma}{Matrix}(
  x,
  uri,
  soma_parent,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  key = NULL,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

\method{write_soma}{TsparseMatrix}(
  x,
  uri,
  soma_parent,
  type = NULL,
  transpose = FALSE,
  ...,
  key = NULL,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)
}
\arguments{
\item{x}{An object}

\item{uri}{URI for resulting SOMA object}

\item{soma_parent}{The parent \link[tiledbsoma:SOMACollection]{collection}
(eg. a \code{\link{SOMACollection}}, \code{\link{SOMAExperiment}}, or
\code{\link{SOMAMeasurement}})}

\item{df_index}{The name of the column in \code{x} with the index
(row names); by default, will automatically add the row names of \code{x}
to an attribute named \dQuote{\code{index}} to the resulting
\code{\link{SOMADataFrame}}}

\item{index_column_names}{Names of columns in \code{x} to index in the
resulting SOMA object}

\item{...}{Arguments passed to other methods}

\item{platform_config}{Optional \link[tiledbsoma:PlatformConfig]{platform
configuration}}

\item{tiledbsoma_ctx}{Optional \code{\link{SOMATileDBContext}}}

\item{relative}{\strong{[Internal use only]} Is \code{uri}
relative or aboslute}

\item{sparse}{Create a \link[tiledbsoma:SOMASparseNDArray]{sparse} or
\link[tiledbsoma:SOMADenseNDArray]{dense} array from \code{x}}

\item{type}{\link[arrow:data-type]{Arrow type} for encoding \code{x}
(eg. \code{\link[arrow:data-type]{arrow::int32}()}); by default, attempts to
determine arrow type with \code{\link[arrow:infer_type]{arrow::infer_type}()}}

\item{transpose}{Transpose \code{x} before writing}

\item{key}{Optionally register the resulting \code{SOMADataFrame} in
\code{soma_parent} as \code{key}; pass \code{NULL} to prevent registration
to handle manually}
}
\value{
The resulting SOMA \link[tiledbsoma:SOMASparseNDArray]{array} or
\link[tiledbsoma:SOMADataFrame]{data frame}, returned opened for write
}
\description{
Various helpers to write R objects to SOMA
}
\section{Writing Character Arrays}{

\link[base:character]{Characters} are written out as
\code{\link{SOMADataFrame}s} with one attribute: \dQuote{\code{values}};
additionally one bit of array-level metadata is added:
\itemize{
\item \dQuote{\code{\Sexpr[stage=build]{names(tiledbsoma:::uns_hint())}}}
with a value of \dQuote{\code{\Sexpr[stage=build]{tiledbsoma:::uns_hint()[[1L]]}}}
}
}

\section{Writing Data Frames}{

\link[base:data.frame]{Data frames} are written out as
\code{\link{SOMADataFrame}s}. The following transformations
are applied to \code{x}:
\itemize{
\item row names are added to a column in \code{x} entitled
\dQuote{\code{index}}, \dQuote{\code{_index}}, or a random name if
either option is already present in \code{x}
\item a column \dQuote{\code{soma_joinid}} will be automatically
added going from \code{[0, nrow(x) - 1]} encoded as
\link[bit64:integer64]{64-bit integers}
}
The array type for each column will be determined by
\code{\link[arrow:infer_type]{arrow::infer_type}()}; if any column contains
a \link[base:is.atomic]{non-atomic} type (excluding
\link[base:factor]{factors}, \code{\link[base]{complex}es},and
\code{\link[base]{raw}s}), the code will error out
}

\section{Writing Dense Matrices}{

Dense matrices are written as two-dimensional
\link[tiledbsoma:SOMADenseNDArray]{dense arrays}. The overall shape of the
array is determined by \code{dim(x)} and the type of the array is determined
by \code{type} or \code{\link[arrow:infer_type]{arrow::infer_type}(x)}
}

\section{Writing Sparse Matrices}{

Sparse matrices are written out as two-dimensional
\link[tiledbsoma:SOMASparseNDArray]{TileDB sparse arrays} in
\href{https://en.wikipedia.org/wiki/Sparse_matrix#Coordinate_list_(COO)}{COO format}:
\itemize{
\item the row indices (\dQuote{\code{i}}) are written out as
\dQuote{\code{soma_dim_0}}
\item the column indices (\dQuote{\code{j}}) are written out as
\dQuote{\code{soma_dim_1}}
\item the non-zero values (\dQuote{\code{x}}) are written out as
\dQuote{\code{soma_data}}
}
The array type is determined by \code{type}, or
\code{\link[arrow:infer_type]{arrow::infer_type}(slot(x, "x"))}
}

\keyword{internal}
