% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seurat.R
\name{write_soma_seurat_sub_objects}
\alias{write_soma_seurat_sub_objects}
\alias{write_soma.Assay}
\alias{write_soma.Assay5}
\alias{write_soma.DimReduc}
\alias{write_soma.Graph}
\alias{write_soma.SeuratCommand}
\title{Convert a \pkg{Seurat} Sub-Object to a SOMA Object, returned opened for write}
\usage{
\method{write_soma}{Assay}(
  x,
  uri = NULL,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)

\method{write_soma}{Assay5}(
  x,
  uri = NULL,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)

\method{write_soma}{DimReduc}(
  x,
  uri = NULL,
  soma_parent,
  fidx = NULL,
  nfeatures = NULL,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)

\method{write_soma}{Graph}(
  x,
  uri,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)

\method{write_soma}{SeuratCommand}(
  x,
  uri = NULL,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)
}
\arguments{
\item{x}{A \pkg{Seurat} sub-object (eg. an
\code{\link[SeuratObject]{Assay}}, a \code{\link[SeuratObject]{DimReduc}},
or a \code{\link[SeuratObject]{Graph}}).}

\item{uri}{URI for resulting SOMA object.}

\item{soma_parent}{The parent \link[tiledbsoma:SOMACollection]{collection};
for the \code{DimReduc} and \code{Graph} methods, this \strong{must} be a
\link[tiledbsoma:SOMAMeasurement]{measurement} for the assay \code{x}
was generated from.}

\item{...}{Arguments passed to other methods}

\item{ingest_mode}{Ingestion mode when creating the SOMA; choose from:
\itemize{
\item \dQuote{\code{write}}: create a new SOMA and error if it already
exists.
\item \dQuote{\code{resume}}: attempt to create a new SOMA; if it already
exists, simply open it for writing.
}}

\item{platform_config}{Optional \link[tiledbsoma:PlatformConfig]{platform
configuration}.}

\item{tiledbsoma_ctx}{Optional (DEPRECATED) \code{\link{SOMATileDBContext}}.}

\item{context}{Optional \code{SOMAContext} object used for TileDB operations.
If a context is not provided, then the default context will be used. Call
\code{set_default_context} once before other SOMA operations to configure
the default context.}

\item{relative}{\strong{[Internal use only]} Is \code{uri}
relative or absolute.}

\item{fidx}{An integer vector describing the location of features in
\code{SeuratObject::Loadings(x)} with relation to \code{soma_parent}
(eg. \code{match(rownames(Loadings(x)), rownames(assay))}).}

\item{nfeatures}{The number of features present in \code{soma_parent}
(eg. \code{nrow(assay)}).}
}
\value{
\code{Assay} and \code{Assay5} methods: a
\code{\link{SOMAMeasurement}} with the data from \code{x},
returned opened for write.

\code{DimReduc} and \code{Graph} methods: invisibly returns
\code{soma_parent}, opened for write, with the values of \code{x}
added to it.
}
\description{
Various helpers to write \pkg{Seurat} sub-objects to SOMA objects..
}
\section{Writing v3 \code{\link[SeuratObject]{Assay}s}}{

\pkg{Seurat} \code{\link[SeuratObject]{Assay}} objects are written out as
individual \link[tiledbsoma:SOMAMeasurement]{measurements}:
\itemize{
\item the \dQuote{\code{data}} matrix is written out as a
\link[tiledbsoma:SOMASparseNDArray]{sparse array} called
\dQuote{\code{data}} within the \dQuote{\code{X}} group.
\item the \dQuote{\code{counts}} matrix, if not
\link[SeuratObject:IsMatrixEmpty]{empty}, is written out as a
\link[tiledbsoma:SOMASparseNDArray]{sparse array} called
\dQuote{\code{counts}} within the \dQuote{\code{X}} group.
\item the \dQuote{\code{scale.data}} matrix, if not
\link[SeuratObject:IsMatrixEmpty]{empty}, is written out as a
\link[tiledbsoma:SOMASparseNDArray]{sparse array} called
\dQuote{\code{scale_data}} within the \dQuote{\code{X}} group.
\item feature-level metadata is written out as a
\link[tiledbsoma:SOMADataFrame]{data frame} called \dQuote{\code{var}}.
}
Expression matrices are transposed (cells as rows) prior to writing. All
other slots, including results from extended assays (eg. \code{SCTAssay},
\code{ChromatinAssay}) are lost.

\subsection{Performance Considerations}{
Ingestion of very large dense layers, such as \code{scale.data}, can be
memory intensive. For better performance, users can remove these layers
prior to ingestion and regenerate them after export, or ingest them
separately as dense arrays for those who need to persist the exact matrix

\preformatted{
# Using SeuratObject v5 syntax on a v3 `Assay`
# Cache the layer for separate ingestion, skip if planning to regenerate
mat <- object[["ASSAY"]]$scale.data

# Remove the `scale.data` layer
object[["ASSAY"]]$scale.data <- NULL

# Ingest the smaller object
uri <- write_soma(object, "/path/to/soma")

# Ingest the `scale.data` layer densely; needed only if persistence
# of the data is paramount
# Pad the `scale.data` layer so that its soma join IDs match the experiment
padded <- matrix(
  data = vector("numeric", length = prod(dim(object[["ASSAY"]]))),
  nrow = nrow(object[["ASSAY"]]),
  ncol = ncol(object[["ASSAY"]])
)
rowidx <- match(rownames(mat), rownames(object[["ASSAY"]]))
colidx <- match(colnames(mat), colnames(object[["ASSAY"]]))
padded[rowidx, colidx] <- mat

# Use `write_soma()` to ingest densely and register it within the `uns`
# collection; this may need to be created manually if the original
# object does not contain command logs
exp <- SOMAExperimentOpen(uri, "WRITE")
if (!match("uns", exp$names(), nomatch = 0L)) {
  # For `tiledb://` URIs, set the URI for the new collection manually rather
  # than relying on `file.path()`
  uns <- SOMACollectionCreate(file.path(exp$uri, "uns"))
  exp$add_new_collection(uns, "uns")
}
arr <- write_soma(
  padded,
  "scale_data",
  soma_parent = exp$get("uns"),
  sparse = FALSE,
  key = "scale_data"
)
arr$close()
exp$close()
}
Please note that dense arrays cannot be read in using the
\code{\link{SOMAExperimentAxisQuery}} mechanism; use
\code{\link{SOMADenseNDArray}$read_dense_matrix}, remembering to transpose
before adding back to a \code{Seurat} object
}
}

\section{Writing v5 \code{Assays}}{

\pkg{Seurat} v5 \code{\link[SeuratObject:Assay5]{Assays}s} are written
out as individual \link[tiledbsoma:SOMAMeasurement]{measurements}:
\itemize{
\item the layer matrices are written out as
\link[tiledbsoma:SOMASparseNDArray]{sparse arrays} within the
\dQuote{\code{X}} group.
\item feature-level metadata is written out as a
\link[tiledbsoma:SOMADataFrame]{data frame} called \dQuote{\code{var}}.
}
Expression matrices are transposed (cells as rows) prior to writing. All
other slots, including results from extended assays (eg. \code{SCTAssay},
\code{ChromatinAssay}) are lost.\cr
The following bits of metadata are written in various parts of the measurement
\itemize{
\item \dQuote{\code{soma_ecosystem_seurat_assay_version}}: written at the
measurement level; indicates the Seurat assay version.
Set to \dQuote{\code{v5}}.
\item \dQuote{\code{soma_ecosystem_seurat_v5_default_layers}}: written at
the \dQuote{\code{X}} group level; indicates the
\link[SeuratObject:DefaultLayer]{default layers}.
\item \dQuote{\code{soma_ecosystem_seurat_v5_ragged}}: written at the
\dQuote{\code{X/<layer>}} array level; with a value of
\dQuote{\code{ragged}}, indicates whether or not the layer is ragged.
\item \dQuote{\code{soma_r_type_hint}}: written at the
\dQuote{\code{X/<layer>}} array level; indicates the \R class and
defining package (for S4 classes) of the original layer.
}
}

\section{Writing \code{\link[SeuratObject]{DimReduc}s}}{

\pkg{Seurat} \code{\link[SeuratObject]{DimReduc}} objects are written out
to the \dQuote{\code{obsm}} and \dQuote{\code{varm}} groups of a
\link[tiledbsoma:SOMAMeasurement]{measurement}:
\itemize{
\item cell embeddings are written out as a
\link[tiledbsoma:SOMASparseNDArray]{sparse matrix} in the
\dQuote{\code{obsm}} group.
\item feature loadings, if not \link[SeuratObject:IsMatrixEmpty]{empty},
are written out as a \link[tiledbsoma:SOMASparseNDArray]{sparse matrix} in
the \dQuote{\code{varm}} groups; loadings are padded with \code{NAs}
to include all features.
}
Dimensional reduction names are translated to AnnData-style names (eg.
\dQuote{\code{pca}} becomes \code{X_pca} for embeddings and
\dQuote{\code{PCs}} for loadings). All other slots, including projected
feature loadings and jackstraw information, are lost.
}

\section{Writing \code{\link[SeuratObject]{Graph}s}}{

\pkg{Seurat} \code{\link[SeuratObject]{Graph}} objects are
written out as \link[tiledbsoma:SOMASparseNDArray]{sparse matrices}
to the \dQuote{\code{obsp}} group of a
\link[tiledbsoma:SOMAMeasurement]{measurement}.
}

\section{Writing \code{\link[SeuratObject]{SeuratCommand}s}}{

\pkg{Seurat} \link[SeuratObject:SeuratCommand]{command logs} are written out
as \link[tiledbsoma:SOMADataFrame]{data frames} to the
\dQuote{\code{seurat_commands}} group of a
\link[tiledbsoma:SOMACollection]{collection}.
}

\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE) && requireNamespace("SeuratObject", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
uri <- withr::local_tempfile(pattern = "seurat-sub")

data("pbmc_small", package = "SeuratObject")
suppressWarnings(pbmc_small <- SeuratObject::UpdateSeuratObject(pbmc_small))

col <- SOMACollectionCreate(uri)

# Write a v3 Assay
(assay <- pbmc_small[["RNA"]])
(ms <- write_soma(assay, "RNA", soma_parent = col))

# Write a v5 Assay
(assay5 <- methods::as(pbmc_small[["RNA"]], "Assay5"))
(ms5 <- write_soma(assay5, "RNA5", soma_parent = col))

ms5$close()

# Write a dimensional reduction
(tsne <- pbmc_small[["tsne"]])
write_soma(tsne, soma_parent = ms)
ms$obsm

# Write a Seurat Graph
(snn <- pbmc_small[["RNA_snn"]])
write_soma(snn, "snn", soma_parent = ms)
ms$obsp

# Write a Seurat command log
(cmd <- pbmc_small[["NormalizeData.RNA"]])
write_soma(cmd, "NormalizeData.RNA", soma_parent = col)
(logs <- col$get("seurat_commands"))
logs$get("NormalizeData.RNA")

col$close()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
